% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation_core_model.R
\name{aggregate_cohort_to_herd}
\alias{aggregate_cohort_to_herd}
\title{Aggregate Cohort-Level Data to Herd-Level}
\usage{
aggregate_cohort_to_herd(data_cohort, id_cols, vars_to_sum, cohort_short)
}
\arguments{
\item{data_cohort}{Cohort-level dataset containing energy allocation variables and herd identifiers. Each row
corresponds to a single sex–age cohort within a herd.}

\item{id_cols}{Character. Unique identifier for the herd, repeated for each cohort belonging to the same herd.}

\item{vars_to_sum}{Character vector. Names of numeric cohort-level variables to be summed across cohorts to produce
herd-level totals (e.g., energy_allocation_meat, energy_allocation_milk, energy_allocation_fibre,
energy_allocation_work, energy_allocation_eggs)}

\item{cohort_short}{Character. Sex- and age-specific cohort code describing the
production stage of the animals. Supported values include:
\itemize{
\item \code{FA}: adult females (from age at first parturition)
\item \code{FS}: sub-adult females (from weaning to age at first parturition)
\item \code{FJ}: juvenile females (from birth to weaning)
\item \code{MA}: adult males (from age at first breeding)
\item \code{MS}: sub-adult males (from weaning to age at first breeding)
\item \code{MJ}: juvenile males (from birth to weaning)
}}
}
\value{
A \code{data.table} at herd scale, in which selected cohort-level variables have been summed across all cohorts
belonging to the same herd, as defined by id_herd.
}
\description{
This function aggregates a dataset from cohort level to herd level by summing
specified variables over the defined 'id' columns.
}
