% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_pathway.R
\name{aggregate_pathway}
\alias{aggregate_pathway}
\title{Aggregate pathway scores}
\usage{
aggregate_pathway(
  score,
  by,
  fun = c("mean", "median", "fraction"),
  threshold = 0,
  long = TRUE
)
}
\arguments{
\item{score}{\code{scpathway_score} object.}

\item{by}{Grouping variable(s): metadata column name(s), or vector.}

\item{fun}{Aggregation function: mean, median, or fraction.}

\item{threshold}{Threshold used by \code{fraction}.}

\item{long}{Return long-format table if TRUE.}
}
\value{
Aggregated table (long) or pathway-by-group matrix (wide).
}
\description{
Aggregate pathway scores
}
