% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_plot.R
\name{apply_gleam_theme}
\alias{apply_gleam_theme}
\title{Apply GLEAM theme to a ggplot}
\usage{
apply_gleam_theme(p, ...)
}
\arguments{
\item{p}{A ggplot object.}

\item{...}{Parameters passed to \code{\link[=gleam_theme]{gleam_theme()}}.}
}
\value{
A ggplot object.
}
\description{
Apply GLEAM theme to a ggplot
}
