% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_model_emissions_ration.R
\name{calc_co2_ration_fertilizer}
\alias{calc_co2_ration_fertilizer}
\title{Calculate a ration component's contribution to carbon dioxide (CO2) emissions from fertilizer manufacture}
\usage{
calc_co2_ration_fertilizer(feed_ration_fraction, co2_feed_fertilizer)
}
\arguments{
\item{feed_ration_fraction}{Numeric. Proportion of a specific feed component in
the total ration,as its fraction of diet dry matter intake (fraction). Within
each herd_id and cohort, proportions should sum to 1.}

\item{co2_feed_fertilizer}{Numeric. Carbon dioxide (CO2) emission factor of a
feed component, representing CO2 emissions from fertilizer manufacture in feed
production, expressed per kilogram of dry matter intake (g CO2/kg DM).}
}
\value{
Numeric. Contribution of an individual feed component to the diet-level average
carbon dioxide (CO2) emission factor from fertilizer manufacture in feed production
(g CO2/kg DM).
}
\description{
Calculates the contribution of an individual feed component to carbon dioxide (CO2)
emissions from fertilizer manufacture in feed production,
using feed-specific emission factors weighted by the component's share in the ration.
}
\details{
The contribution is computed as:

\deqn{diet\_co2\_feed\_fertilizer = feed\_ration\_fraction \times co2\_feed\_fertilizer}

This function is part of the \code{\link[=run_emissions_ration_module]{run_emissions_ration_module()}}.
}
\seealso{
\code{\link{run_emissions_ration_module}}
}
