% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_model_emissions_ration.R
\name{calc_co2_ration_luc_peat}
\alias{calc_co2_ration_luc_peat}
\title{Calculate a ration component's contribution to carbon dioxide (CO2) emissions from peatland drainage}
\usage{
calc_co2_ration_luc_peat(feed_ration_fraction, co2_feed_luc_peat)
}
\arguments{
\item{feed_ration_fraction}{Numeric. Proportion of a specific feed component in the total
ration, expressed as its fraction of diet dry matter intake (fraction). Within
each herd_id and cohort, proportions should sum to 1.}

\item{co2_feed_luc_peat}{Numeric. Carbon dioxide (CO2) emission factor of a feed component,
representing CO2 emissions from peatland drainage in feed production,
expressed per kilogram of dry matter intake (g CO2/kg DM).}
}
\value{
Numeric. Contribution of an individual feed component to the diet-level
average carbon dioxide (CO2) emission factor from  peatland drainage in feed
production (g CO2/kg DM).
}
\description{
Computes the contribution of an individual feed component to carbon dioxide (CO2)
emissions from peatland drainage in feed production,
using feed-specific emission factors weighted by the component's share in the ration.
}
\details{
The contribution is computed as:

\deqn{diet\_co2\_feed\_luc\_peat = feed\_ration\_fraction \times co2\_feed\_luc\_peat}
}
