% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_model_emissions_enteric.R
\name{calc_conversion_factor_ym}
\alias{calc_conversion_factor_ym}
\title{Calculate methane conversion factor (ym)}
\usage{
calc_conversion_factor_ym(
  species_short,
  cohort_short,
  ration_digestibility_fraction
)
}
\arguments{
\item{species_short}{Character. Code identifying the livestock species.
Supported values include:
\itemize{
\item \code{PGS}: pigs
\item \code{CML}: camels
\item \code{CTL}: cattle
\item \code{BFL}: buffalo
\item \code{SHP}: sheep
\item \code{GTS}: goats
\item \code{CHK}: chickens
}}

\item{cohort_short}{Character. Sex- and age-specific cohort code describing the
production stage of the animals. Supported values include:
\itemize{
\item \code{FA}: adult females (from age at first parturition)
\item \code{FS}: sub-adult females (from weaning to age at first parturition)
\item \code{FJ}: juvenile females (from birth to weaning)
\item \code{FN}: non-demographic females
\item \code{MA}: adult males (from age at first breeding)
\item \code{MS}: sub-adult males (from weaning to age at first breeding)
\item \code{MJ}: juvenile males (from birth to weaning)
\item \code{MN}: non-demographic males
}}

\item{ration_digestibility_fraction}{Numeric. Average digestibility of the feed ration, expressed as ratio of
digestible (or metabolizable, for poultry) to gross energy content (fraction).}
}
\value{
Numeric. Methane (CH4) conversion factor (ym), representing the percentage of  gross energy
of the feed ration that is converted to CH4 (percentage).
}
\description{
Calculates the methane conversion factor (ym, \% of dietary gross energy in feed converted to methane)
for a given species and cohort based on diet digestibility. Implements species- and cohort-specific
rules consistent with IPCC Tier 2 approach.
}
\details{
ym is computed using species- and cohort-specific default relationships with diet digestibility (Opio et al., 2013).

\code{ration_digestibility_fraction} can be calculated with
\code{\link{calc_ration_digestibility}} - see also \code{\link{run_ration_quality_module}}.

\itemize{
\item \strong{For \code{CTL} and \code{BFL}:}
\deqn{ym = 9.75 - 0.05 \times (ration\_digestibility\_fraction \times 100)}

\item \strong{For \code{SHP}, \code{GTS} and \code{CML}:}
\itemize{
\item \code{FA} and \code{MA} cohorts:
\deqn{ym = 9.75 - 0.05 \times (ration\_digestibility\_fraction \times 100)}
\item \code{FS}, \code{MS}, \code{FN}, and \code{MN} cohorts:
\deqn{ym = 7.75 - 0.05 \times (ration\_digestibility\_fraction \times 100)}
}

\item \strong{For \code{PGS}:}
ym is assigned fixed values by cohort:
\itemize{
\item \code{FA} and \code{MA} cohorts: \deqn{ym = 1.01}
\item \code{FS}, \code{MS}, \code{FN}, and \code{MN} cohorts: \deqn{ym = 0.39}
}
\item \strong{For \code{CHK}}: ym is returned as 0

}

ym is returned as 0 for juvenile cohorts (\code{FJ}, \code{MJ}), assuming
negligible enteric methane production before weaning/rumen development.

This function is part of the \code{\link[=run_emissions_enteric_module]{run_emissions_enteric_module()}}.
}
\references{
Opio, C., Gerber, P., Mottet, A., Falcucci, A., Tempio, G.,
MacLeod, M., Vellinga, T., Henderson, B. & Steinfeld, H. (2013).
\emph{Greenhouse gas emissions from ruminant supply chains – A global life cycle assessment}.
Food and Agriculture Organization of the United Nations (FAO), Rome.

Jørgensen, H., Theil, P. K. & Knudsen, K. E. B. (2011).
\emph{Enteric methane emission from pigs}.
In: Planet Earth 2011 – Global Warming Challenges and Opportunities for Policy and Practice
(p. 610; Table 2). InTech.

IPCC. (2019).
\emph{2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories}.
Chapter 10: Emissions from Livestock and Manure Management, Equation 10.21.

IPCC. (2006).
\emph{2006 IPCC Guidelines for National Greenhouse Gas Inventories}.
Chapter 10: Emissions from Livestock and Manure Management, Equation 10.21.
}
\seealso{
\code{\link{run_emissions_enteric_module}},
\code{\link{calc_ration_digestibility}},
\code{\link{run_ration_quality_module}}
}
