% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_simulation_functions.R
\name{calc_daily_gain}
\alias{calc_daily_gain}
\title{Calculate Daily Weight Gain}
\usage{
calc_daily_gain(potfinalLW, initialLW, duration)
}
\arguments{
\item{potfinalLW}{Numeric. Potential final live weight.}

\item{initialLW}{Numeric. Initial live weight.}

\item{duration}{Numeric. Duration of the stage (in days).}
}
\value{
Numeric. Daily weight gain (kg/day).
}
\description{
Computes average daily weight gain over a given duration based on the difference
between potential final and initial live weights.
}
