% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_core_model.R
\name{calc_daily_weight_gain}
\alias{calc_daily_weight_gain}
\title{Calculate Daily Weight Gain}
\usage{
calc_daily_weight_gain(potential_final_weight, initial_weight, duration)
}
\arguments{
\item{potential_final_weight}{Numeric. Potential final live weight attainable at the end of the cohort stage in the absence of offtake (kg). (For juveniles: equals weaning weight; For subadults: equals adult live weight; For adults: equals adult live weight)}

\item{initial_weight}{Numeric. Live weight at the beginning of the cohort stage (kg).}

\item{duration}{Numeric. Amount of time that each animal spends in a specific cohort (days).}
}
\value{
Numeric. Average live weight gain of the cohort over the cohort stage (kg/head/day).
}
\description{
Computes average daily weight gain over a given duration based on the difference
between potential final and initial live weights.
}
\details{
Daily live weight gain is calculated as the difference between the
potential final live weight and the initial live weight, divided by the
duration of the cohort stage:

\deqn{daily\_weight\_gain = (potential\_final\_weight - initial\_weight) / duration}

This represents an average (per-head) daily gain.
}
