% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirect_emissions_feed_core_model.R
\name{calc_diet_co2_feed_fertilizer}
\alias{calc_diet_co2_feed_fertilizer}
\title{Calculate diet Carbon dioxide (CO₂) from fertilizer used for feed production}
\usage{
calc_diet_co2_feed_fertilizer(feed_ration_fraction, co2_feed_fertilizer)
}
\arguments{
\item{feed_ration_fraction}{Numeric. Proportion of a specific feed item in the total ration, expressed as its fraction of diet dry matter (fraction). Within each herd_id and cohort, proportions should sum to 1.}

\item{co2_feed_fertilizer}{Numeric. Carbon dioxide (CO₂) emission factor from fertilizer use in feed production, calculated per kg of dry matter intake (g CO₂/kg DM).}
}
\value{
Numeric. Average carbon dioxide (CO₂) emission factor from fertilizer use in feed production of the diet (g CO₂/kg DM).
}
\description{
Computes the weighted contribution of Carbon dioxide (CO₂) emissions from fertilizer use
associated with feed production for a feed ration component.
}
\details{
The contribution is computed as:

\deqn{diet\_co2\_feed\_fertilizer = feed\_ration\_fraction \times co2\_feed\_fertilizer}
}
