% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirect_emissions_feed_core_model.R
\name{calc_diet_co2_feed_pesticides}
\alias{calc_diet_co2_feed_pesticides}
\title{Calculate diet CO2 from pesticide use (feed production)}
\usage{
calc_diet_co2_feed_pesticides(feed_ration_fraction, co2_feed_pesticides)
}
\arguments{
\item{feed_ration_fraction}{Numeric. Fraction of the total ration represented
by this feed component (unitless).}

\item{co2_feed_pesticides}{Numeric. Emission factor for CO2 from pesticide use
in feed production (g CO2/kg dry matter of feed).}
}
\value{
Numeric. CO2 contribution from pesticides for this feed item (g CO2/kg dry matter of feed).
}
\description{
Computes the weighted contribution of CO2 emissions from pesticide use
associated with feed production for a ration component.
}
\details{
\deqn{diet\_co2\_feed\_pesticides = feed\_ration\_fraction \times co2\_feed\_pesticides}

This helper is vectorized over its inputs.
}
