% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_requirements_core_model.R
\name{calc_dry_matter_intake}
\alias{calc_dry_matter_intake}
\title{Calculate Daily Dry Matter Intake}
\usage{
calc_dry_matter_intake(animal, total_energy, diet_ge, diet_me)
}
\arguments{
\item{animal}{Character. Species code.}

\item{total_energy}{Numeric. Total energy requirement (MJ/head/day).}

\item{diet_ge}{Numeric. Gross energy of diet (MJ/kg DM).}

\item{diet_me}{Numeric. Metabolizable energy of diet (MJ/kg DM).}
}
\value{
Numeric. Dry matter intake (kg DM/head/day).
}
\description{
Computes daily feed intake per animal (kg DM/head/day).
}
