% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_model_ration_quality.R
\name{calc_feed_digestibility_fraction}
\alias{calc_feed_digestibility_fraction}
\title{Calculate feed digestibility fraction}
\usage{
calc_feed_digestibility_fraction(
  feed_digestible_energy_ruminant = NA_real_,
  feed_digestible_energy_pigs = NA_real_,
  feed_gross_energy
)
}
\arguments{
\item{feed_digestible_energy_ruminant}{Numeric. Digestible energy content of a
feed component for ruminants, representing the energy absorbed by the animal after
fecal losses (MJ/kg DM).}

\item{feed_digestible_energy_pigs}{Numeric. Digestible energy content of a feed
component for pigs, representing the energy absorbed by the animal after fecal
losses (MJ/kg DM).}

\item{feed_gross_energy}{Numeric. Gross energy content of a feed component,
representing the total chemical energy released upon complete combustion of
the feed (MJ/kg DM).}
}
\value{
List with elements:
\describe{
\item{feed_digestibility_fraction_ruminant}{
Numeric. Digestibility of a feed component for ruminants, expressed as the ratio
of digestible energy to gross energy content (fraction).
}
\item{feed_digestibility_fraction_pigs}{
Numeric. Digestibility of a feed component for pigs, expressed as the ratio of
digestible energy to gross energy content (fraction).
}
}
}
\description{
Calculates species-specific feed digestibility fractions by feed component.
}
\details{
Digestibility is computed as the ratio of usable energy to gross energy:
\deqn{feed\_digestibility\_fraction = usable\_energy / feed\_gross\_energy}

For ruminants and pigs, usable energy is represented by \code{digestible_energy} (DE),
which accounts for fecal energy losses.

This function is part of the \code{\link[=run_ration_quality_module]{run_ration_quality_module()}}.
}
\seealso{
\code{\link{run_ration_quality_module}}
}
