% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_model_metabolic_energy_req.R
\name{calc_metabolic_energy_req_fibre}
\alias{calc_metabolic_energy_req_fibre}
\title{Calculate metabolic energy requirements for fibre production}
\usage{
calc_metabolic_energy_req_fibre(
  species_short,
  cohort_short,
  fibre_yield_year = NA_real_
)
}
\arguments{
\item{species_short}{Character. Code identifying the livestock species.
Supported values include:
\itemize{
\item \code{PGS}: pigs
\item \code{CML}: camels
\item \code{CTL}: cattle
\item \code{BFL}: buffalo
\item \code{SHP}: sheep
\item \code{GTS}: goats
\item \code{CHK}: chickens
}}

\item{cohort_short}{Character. Sex- and age-specific cohort code describing the production stage of the animals. Supported values include:
\itemize{
\item \code{FA}: adult females (from age at first parturition)
\item \code{FS}: sub-adult females (from weaning to age at first parturition)
\item \code{FJ}: juvenile females (from birth to weaning)
\item \code{FN}: non-demographic females
\item \code{MA}: adult males (from age at first breeding)
\item \code{MS}: sub-adult males (from weaning to age at first breeding)
\item \code{MJ}: juvenile males (from birth to weaning)
\item \code{MN}: non-demographic males
}}

\item{fibre_yield_year}{Numeric. Annual production yield of fibre, such as wool, cashmere, mohair (kg/head/year).
Required only for species = CML, SHP, and GTS.}
}
\value{
Numeric. Energy required for the synthesis of fibre for SHP, GTS and CML.
Assumed to be 0 for other species. (MJ/head/day).  Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).
}
\description{
Calculates the energy requirement for fibre production (MJ/head/day),
defined as the additional energy required for the synthesis of animal fibre
(e.g. wool or hair).
}
\details{
This component follows the IPCC Tier 2 partitioning approach and is applied
only to fibre-producing species and relevant cohorts, which are assumed to be
\code{FA}, \code{FS}, \code{MA}, \code{MS}, \code{FN}, and \code{MN}.

\strong{Species-specific approach:}

\itemize{

\item \strong{SHP and GTS} (IPCC, 2006; IPCC, 2019):

For sheep and goats, fibre production energy is calculated assuming a fixed
net energy cost of \eqn{24} MJ per kilogram of fibre produced. Annual fibre
production is converted to a daily requirement as:

\deqn{
metabolic\_energy\_req\_fibre =
\frac{24 \times fibre\_yield\_year}{365}
}

where:
\itemize{
\item \eqn{fibre\_yield\_year} is annual fibre production
(kg/head/year),
\item \eqn{24} is the net energy requirement per kilogram of fibre
(MJ/kg fibre),
\item division by \eqn{365} converts annual production to a daily basis.
}

\item \strong{CML} (AFRC, 1998; Cannas et al., 2007):

For camels, fibre energy requirements are first calculated on a
\strong{net energy} basis and then converted to
\strong{metabolizable energy} using a net-to-metabolizable energy
efficiency coefficient:

\deqn{
metabolic\_energy\_req\_fibre =
\frac{24}{0.43}
\times
\frac{fibre\_yield\_year}{365}
}

where:
\itemize{
\item \eqn{24} is the net energy requirement per kilogram of fibre
(MJ/kg fibre),
\item \eqn{0.43} is the efficiency of conversion from metabolizable energy
to net energy for fibre production (fraction),
\item \eqn{fibre\_prod} is annual fibre production per animal
(kg/head/year),
\item \eqn{365} to convert annual production to a daily basis.
}

The efficiency coefficient of \eqn{0.43} is adopted by analogy with goats,
assuming a dietary metabolizability of approximately 0.55, following AFRC
guidance and the synthesis by Cannas et al. (2007).

\item\strong{Other species}: Fibre production energy is assumed to be zero.
}

This function is part of the \code{\link[=run_metabolic_energy_req_module]{run_metabolic_energy_req_module()}}.
}
\references{
AFRC (1998) \emph{The Nutrition of Goats.}
CAB International, Wallingford, UK.

AFRC (1993). \emph{Energy and Protein Requirements of Ruminants. An Advisory Manual Prepared by the AFRC Technical Committee on Responses to Nutrients.}
CAB International, Wallingford, UK.

Cannas, A., Atzori, A. S., Boe, F., & Teixeira, I. (2007).
\emph{Energy and protein requirements of goats}.
In: Dairy sheep nutrition (pp. 31-49). CAB International, Wallingford, UK.

IPCC. (2019). \emph{2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories}.
Chapter 10: Emissions from Livestock and Manure Management, Equation 10.12.

IPCC. (2006). \emph{2006 IPCC Guidelines for National Greenhouse Gas Inventories}. Chapter 10: Emissions from
Livestock and Manure Management, Equation 10.12.
}
\seealso{
\code{\link{run_metabolic_energy_req_module}},
}
