% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_requirements_core_model.R
\name{calc_net_energy_activity}
\alias{calc_net_energy_activity}
\title{Calculate Net Energy for Activity}
\usage{
calc_net_energy_activity(
  animal,
  cohort,
  past_man_frac,
  mmspasture,
  nemain,
  average_weight,
  offtake_rate
)
}
\arguments{
\item{animal}{Character. Species code.}

\item{cohort}{Character. Cohort code.}

\item{past_man_frac}{Numeric. Pasture management fraction.}

\item{mmspasture}{Numeric. Fraction of time on pasture.}

\item{nemain}{Numeric. Net energy for maintenance.}

\item{average_weight}{Numeric. Average live weight (kg).}

\item{offtake_rate}{Numeric. Offtake rate by cohort.}
}
\value{
Numeric. Net energy for activity (MJ/head/day).
}
\description{
Computes net energy required for activity (MJ/head/day).
}
