% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_requirements_core_model.R
\name{calc_net_energy_growth}
\alias{calc_net_energy_growth}
\title{Calculate Net Energy for Growth}
\usage{
calc_net_energy_growth(
  animal,
  cohort,
  average_weight,
  final_weight,
  initial_weight,
  dwg,
  offtake_rate,
  duration
)
}
\arguments{
\item{animal}{Character. Species code.}

\item{cohort}{Character. Cohort code.}

\item{average_weight}{Numeric. Average live weight (kg).}

\item{final_weight}{Numeric. Final live weight (kg).}

\item{initial_weight}{Numeric. Initial live weight (kg).}

\item{dwg}{Numeric. Daily weight gain (kg/day).}

\item{offtake_rate}{Numeric. Offtake rate by cohort.}

\item{duration}{Numeric. Duration in days.}
}
\value{
Numeric. Net energy for growth (MJ/head/day).
}
\description{
Computes net energy required for growth (MJ/head/day).
}
