% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_requirements_core_model.R
\name{calc_net_energy_maintenance}
\alias{calc_net_energy_maintenance}
\title{Calculate Net Energy for Maintenance}
\usage{
calc_net_energy_maintenance(
  animal,
  cohort,
  average_weight,
  idle = NA_real_,
  gest = NA_real_,
  lact = NA_real_,
  litsize = NA_real_,
  ckg = NA_real_,
  milking_fraction = NA_real_,
  offtake_rate = NA_real_,
  afc = NA_real_
)
}
\arguments{
\item{animal}{Character. Species code (e.g., "CTL", "BFL", "SHP", "GTS", "PGS", "CHK", "CML").}

\item{cohort}{Character. Cohort code.}

\item{average_weight}{Numeric. Average live weight (kg).}

\item{idle}{Numeric. Fraction of time idle (only for PGS–FA, default NA).}

\item{gest}{Numeric. Fraction of time gestating (only for PGS–FA, default NA).}

\item{lact}{Numeric. Fraction of time lactating (only for PGS–FA, default NA).}

\item{litsize}{Numeric. Litter size (only for PGS–FA, default NA).}

\item{ckg}{Numeric. Birth weight (only for PGS–FA, default NA).}

\item{milking_fraction}{Numeric. Proportion of lactating adult females (only for CTL/BFL–FA, default NA).}

\item{offtake_rate}{Numeric. Offtake rate by cohort (only for some cohorts, default NA).}

\item{afc}{Numeric. Age at first calving (only for SHP, default NA).}
}
\value{
Numeric. Net energy for maintenance (MJ/head/day).
}
\description{
Computes net energy required for maintenance (MJ/head/day).
}
