% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_requirements_core_model.R
\name{calc_net_energy_meat}
\alias{calc_net_energy_meat}
\title{Calculate Net Energy for Meat Production}
\usage{
calc_net_energy_meat(
  animal,
  cohort,
  ckg,
  afc,
  slaughter_weight,
  initial_weight
)
}
\arguments{
\item{animal}{Character. Species code.}

\item{cohort}{Character. Cohort code.}

\item{ckg}{Numeric. Birth weight (kg).}

\item{afc}{Numeric. Age at first calving (for SHP, GTS).}

\item{slaughter_weight}{Numeric. Slaughter weight (kg).}

\item{initial_weight}{Numeric. Initial live weight (kg).}
}
\value{
Numeric. Net energy for meat production (MJ/head).
}
\description{
Computes the energy requirement for meat production (MJ/head).
}
