% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_requirements_core_model.R
\name{calc_net_energy_pregnancy}
\alias{calc_net_energy_pregnancy}
\title{Calculate Energy for Pregnancy}
\usage{
calc_net_energy_pregnancy(
  species_short,
  cohort_short,
  energy_requirement_maintenance = NA_real_,
  parturition_rate = NA_real_,
  litter_size = NA_real_,
  pregnancy_duration = NA_real_,
  non_productive_duration = NA_real_,
  lactation_duration = NA_real_,
  cohort_duration_days = NA_real_,
  offtake_rate = NA_real_
)
}
\arguments{
\item{species_short}{Character. Code identifying the livestock species.
Supported values include:
\itemize{
\item \code{PGS}: pigs
\item \code{CML}: camels
\item \code{CTL}: cattle
\item \code{BFL}: buffalo
\item \code{SHP}: sheep
\item \code{GTS}: goats
}}

\item{cohort_short}{Character. Sex- and age-specific cohort code describing the
production stage of the animals. Supported values include:
\itemize{
\item \code{FA}: adult females (from age at first parturition)
\item \code{FS}: sub-adult females (from weaning to age at first parturition)
\item \code{FJ}: juvenile females (from birth to weaning)
\item \code{MA}: adult males (from age at first breeding)
\item \code{MS}: sub-adult males (from weaning to age at first breeding)
\item \code{MJ}: juvenile males (from birth to weaning)
}}

\item{energy_requirement_maintenance}{Numeric. Energy required for maintenance, defined as the amount of
energy needed to keep the animal in equilibrium such that body energy is neither gained nor lost.
Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).}

\item{parturition_rate}{Numeric. Average annual number of parturitions per female animal (#). A herd-level
reproductive performance indicator calculated as the total number of parturitions (deliveries) occurring
during a year divided by the number of adult females potentially able to give birth during that year.}

\item{litter_size}{Numeric. Average number of offspring born per parturition (#). This value can be
calculated as the total number of offspring born divided by the total number of parturitions during the
year.}

\item{pregnancy_duration}{Numeric. Duration of pregnancy period (days).}

\item{non_productive_duration}{Numeric. Period during which the animal is not performing any productive
physiological function such as pregnancy or lactation (days).}

\item{lactation_duration}{Numeric. Duration of the lactation period, defined as the number of days during
which the animal is lactating (days).}

\item{cohort_duration_days}{Numeric. Amount of time that each animal spends in a specific cohort (days).}

\item{offtake_rate}{Numeric. Annual proportion of animals removed from the herd for each sex-age cohort
(fraction).}
}
\value{
Numeric. Energy required for pregnancy (\code{energy_requirement_pregnancy}) for pregnant adult
females (MJ/head/day). Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML
and PGS (MJ/head/day).
}
\description{
Computes the \strong{energy requirement for pregnancy} (MJ/head/day) for pregnant
females. This approach follows the IPCC Tier 2 partitioning framework and is applied
only to \strong{female cohorts}.
}
\details{
Pregnancy energy is calculated \strong{only for female cohorts}
(i.e., \code{FA} and \code{FS}) and represents the additional
energy required to support gestation.

For \code{FA}, pregnancy energy requirements are adjusted to account
for the proportion of time animals spend in gestation.

For \code{FS}, only a fraction of animals is assumed to be of reproductive age;
pregnancy energy requirements are therefore scaled accordingly.

\itemize{
\item For \strong{CTL and BFL} — IPCC (2006, 2019):

Pregnancy energy is approximated as 10\% of maintenance energy.
\itemize{
\item \code{FA}:
\deqn{
      energy\_pregnancy =
      0.10 \times nemain \times parturition\_rate \times \frac{gest}{365}
      }
\item \code{FS}:
\deqn{
      energy\_pregnancy =
      0.10 \times nemain \times \frac{gest}{duration}
      \times (1 - offtake\_rate)
      }
}

\item For \strong{CML} — Wardeh (2004):

Pregnancy energy is estimated as 12\% of maintenance energy.
\itemize{
\item \code{FA}:
\deqn{
      energy\_pregnancy =
      0.12 \times nemain \times parturition\_rate
      }
\item \code{FS}:
\deqn{
      energy\_pregnancy =
      0.12 \times nemain \times \frac{gest}{duration}
      \times (1 - offtake\_rate)
      }
}

\item For \strong{SHP and GTS} — IPCC (2006, 2019):

Pregnancy energy is calculated as a litter-size–dependent fraction
of maintenance energy.
\itemize{
\item \code{FA}:
\deqn{
      energy\_pregnancy =
      nemain \times cpreg \times parturition\_rate \times \frac{gest}{365}
      }
where \eqn{cpreg} is defined as:
\itemize{
\item If \eqn{1 \le litsize \le 2}:
\deqn{
          cpreg = 0.077 \times (2 - litsize) + 0.126 \times (litsize - 1)
          }
\item If \eqn{litsize > 2}:
\deqn{cpreg = 0.150}
}
\item \code{FS} -
Pregnancy energy is calculated using the single-birth coefficient
and scaled to the proportion of reproductive individuals in the cohort.
\deqn{
      energy\_pregnancy =
      0.077 \times nemain \times \frac{gest}{duration}
      \times (1 - offtake\_rate)
      }
}

\item For \strong{PGS} — NRC (1998):

\itemize{
\item \code{FA}:
\deqn{
      energy\_pregnancy = cgest \times litsize \times \frac{gest}{lact + gest + idle}
      }
\item \code{FS}:
\deqn{
      energy\_pregnancy =
      cgest \times litsize \times \frac{gest}{duration}
      \times (1 - offtake\_rate)
      }
The default value for the gestation coefficient is
\eqn{cgest = 0.14985} MJ piglet\eqn{^{-1}}.
}
}
}
\references{
NRC (1998). \emph{Nutrient Requirements of Swine},
10th Revised Edition. National Academies Press, Washington, DC.

IPCC. (2019). \emph{2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories}.
Chapter 10: Emissions from Livestock and Manure Management, Equation 10.13; Table 10.7.

IPCC. (2006). \emph{2006 IPCC Guidelines for National Greenhouse Gas Inventories}. Chapter 10: Emissions
from Livestock and Manure Management, Equation 10.13; Table 10.7.

Wardeh, M. F. (2004). \emph{The nutrient requirements of the dromedary camel}.
Journal of Camel Science. 2004;1:37–45. The Camel Applied Research and Development Network (CARDN),
Arab Center for the Studies of Arid Zones and Dry Lands (ACSAD).
}
