% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_requirements_core_model.R
\name{calc_net_energy_work}
\alias{calc_net_energy_work}
\title{Calculate Net Energy for Work}
\usage{
calc_net_energy_work(animal, cohort, nemain, work_hours, draught_fraction)
}
\arguments{
\item{animal}{Character. Species code.}

\item{cohort}{Character. Cohort code.}

\item{nemain}{Numeric. Net energy for maintenance.}

\item{work_hours}{Numeric. Number of work hours.}

\item{draught_fraction}{Numeric. Draught fraction.}
}
\value{
Numeric. Net energy for work (MJ/head/day).
}
\description{
Computes net energy required for work (MJ/head/day).
}
