% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directemissions_manure_core_model.R
\name{calc_nitrogen_leaching_fraction}
\alias{calc_nitrogen_leaching_fraction}
\title{Calculate Nitrogen Leaching Fraction}
\usage{
calc_nitrogen_leaching_fraction(
  mms_pasture,
  mms_burned,
  mms_other,
  ef_fracleach_pasture,
  ef_fracleach_burned,
  ef_fracleach_other
)
}
\arguments{
\item{mms_pasture}{Numeric vector of manure management system fraction for pasture (0-1)}

\item{mms_burned}{Numeric vector of manure management system fraction for burned (0-1)}

\item{mms_other}{Numeric vector of manure management system fraction for other systems (0-1)}

\item{ef_fracleach_pasture}{Numeric vector of leaching fraction for pasture (0-1)}

\item{ef_fracleach_burned}{Numeric vector of leaching fraction for burned (0-1)}

\item{ef_fracleach_other}{Numeric vector of leaching fraction for other systems (0-1)}
}
\value{
A named list with:
\describe{
\item{fracleach_pasture}{Leaching fraction for pasture (dimensionless)}
\item{fracleach_burned}{Leaching fraction for burned (dimensionless)}
\item{fracleach_other}{Leaching fraction for other systems (dimensionless)}
}
}
\description{
Calculates the fraction of nitrogen lost via leaching from different
manure management systems.
}
