% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directemissions_manure_core_model.R
\name{calc_nitrogen_volatilization_fraction}
\alias{calc_nitrogen_volatilization_fraction}
\title{Calculate Nitrogen Volatilization Fraction}
\usage{
calc_nitrogen_volatilization_fraction(
  mms_pasture,
  mms_burned,
  mms_other,
  ef_fracgas_pasture,
  ef_fracgas_burned,
  ef_fracgas_other
)
}
\arguments{
\item{mms_pasture}{Numeric vector of manure management system fraction for pasture (0-1)}

\item{mms_burned}{Numeric vector of manure management system fraction for burned (0-1)}

\item{mms_other}{Numeric vector of manure management system fraction for other systems (0-1)}

\item{ef_fracgas_pasture}{Numeric vector of volatilization fraction for pasture (0-1)}

\item{ef_fracgas_burned}{Numeric vector of volatilization fraction for burned (0-1)}

\item{ef_fracgas_other}{Numeric vector of volatilization fraction for other systems (0-1)}
}
\value{
A named list with:
\describe{
\item{fracgas_pasture}{Volatilization fraction for pasture (dimensionless)}
\item{fracgas_burned}{Volatilization fraction for burned (dimensionless)}
\item{fracgas_other}{Volatilization fraction for other systems (dimensionless)}
}
}
\description{
Calculates the fraction of nitrogen lost via volatilization from different
manure management systems.
}
