% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_nondemo_core_model.R
\name{calc_nondemo_start_sizes}
\alias{calc_nondemo_start_sizes}
\title{Allocate Non-Demographic Entrants Across Production Cycles}
\usage{
calc_nondemo_start_sizes(
  cohort_stock_annual_non_demographic,
  total_cycle_starts_to_distribute
)
}
\arguments{
\item{cohort_stock_annual_non_demographic}{Numeric vector. Total number of animals entering the non-demographic model over the simulated period
(heads / simulated period). The simulated period is set by default to 365 days to align with the demographic herd model.}

\item{total_cycle_starts_to_distribute}{Integer vector. Total number of cycle starts to be distributed within the simulated_days window for the non-demographic cohort assessed (FN or MN) (cycle starts/simulated days).}
}
\value{
A named list with the following element:
\describe{
\item{cohort_stock_start_cycle_non_demographic}{Numeric vector. Number of animals starting each non-demographic production cycle for the assessed cohort (FN or MN),
calculated by distributing the total cohort stock over the total number of cycle starts (heads / cycle).}
}
}
\description{
Converts the total annual number of animals entering a non-demographic cohort
into the number of animals entering each production cycle, based on the number
of cycles occurring within the simulated period.
}
\details{
The annual cohort stock is evenly distributed across all production cycles.

If \code{total_cycle_starts_to_distribute} is less than or equal to zero, the function
assumes that no cycle structure is defined and assigns all entrants to a single
cycle.
}
