% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_model_metabolic_energy_req.R
\name{calc_ration_intake}
\alias{calc_ration_intake}
\title{Calculate daily ration intake in dry matter}
\usage{
calc_ration_intake(
  species_short,
  metabolic_energy_req_total,
  ration_gross_energy,
  ration_metabolizable_energy
)
}
\arguments{
\item{species_short}{Character. Code identifying the livestock species.
Supported values include:
\itemize{
\item \code{PGS}: pigs
\item \code{CML}: camels
\item \code{CTL}: cattle
\item \code{BFL}: buffalo
\item \code{SHP}: sheep
\item \code{GTS}: goats
}}

\item{metabolic_energy_req_total}{Numeric. Total daily energy requirement (MJ/head/day).
For CTL, BFL, SHP and GTS this is expressed as gross energy intake requirement (GE).
For CML and PGS the function returns the summed daily metabolizable energy requirement.}

\item{ration_gross_energy}{Numeric. Average gross energy content of the diet (MJ/kg DM).}

\item{ration_metabolizable_energy}{Numeric. Average metabolizable energy content of the diet (MJ/kg DM).}
}
\value{
Numeric. Average daily dry matter intake of feed (kg DM/head/day).
}
\description{
Calculates daily feed intake as dry matter intake (DMI) per animal (kg DM/head/day) from the
animal's daily energy requirement and the diet energy density.
}
\details{
This function follows the IPCC Tier 2 framework.
DMI is computed
by dividing the appropriate daily energy requirement by the corresponding
diet energy content (MJ/kg DM).

\itemize{
\item \strong{Energy expressed as gross energy intake requirement -
\code{CTL}, \code{BFL}, \code{SHP}, \code{GTS}}:

\deqn{
    ration\_intake = \frac{metabolic\_energy\_req\_total}{ration\_gross\_energy}
  }

\item \strong{Energy expressed as metabolizable energy requirement -
\code{CML}, \code{PGS}}:

\deqn{
    ration\_intake = \frac{metabolic\_energy\_req\_total}{ration\_metabolizable\_energy}
  }
}

This function is part of the \code{\link[=run_metabolic_energy_req_module]{run_metabolic_energy_req_module()}}.
}
\references{
IPCC. (2006). \emph{2006 IPCC Guidelines for National Greenhouse Gas Inventories},
Volume 4 (AFOLU), Chapter 10: \emph{Emissions from Livestock and Manure Management}.

IPCC. (2019). \emph{2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories},
Volume 4 (AFOLU), Chapter 10: \emph{Emissions from Livestock and Manure Management}.
}
\seealso{
\code{\link{run_metabolic_energy_req_module}},
\code{\link{calc_total_metabolic_energy_req}}
\code{\link{calc_ration_gross_energy}}
\code{\link{calc_ration_metabolizable_energy}}
}
