% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_requirements_core_model.R
\name{calc_total_energy_requirement}
\alias{calc_total_energy_requirement}
\title{Calculate Total Energy Requirement}
\usage{
calc_total_energy_requirement(
  animal,
  cohort,
  nemain,
  neact,
  nelact,
  nework,
  nepreg,
  rem,
  negrow,
  nefibre,
  neegg,
  reg,
  diet_dig,
  afc
)
}
\arguments{
\item{animal}{Character. Species code.}

\item{cohort}{Character. Cohort code.}

\item{nemain}{Numeric. Net energy for maintenance.}

\item{neact}{Numeric. Net energy for activity.}

\item{nelact}{Numeric. Net energy for lactation.}

\item{nework}{Numeric. Net energy for work.}

\item{nepreg}{Numeric. Net energy for pregnancy.}

\item{rem}{Numeric. REM value.}

\item{negrow}{Numeric. Net energy for growth.}

\item{nefibre}{Numeric. Net energy for fibre production.}

\item{neegg}{Numeric. Net energy for egg production.}

\item{reg}{Numeric. REG value.}

\item{diet_dig}{Numeric. Diet digestibility (fraction).}

\item{afc}{Numeric. Age at first calving (for SHP, GTS).}
}
\value{
Numeric. Total energy requirement (MJ/head/day).
}
\description{
Computes the total energy requirement (MJ/head/day).
}
