% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_model_allocation.R
\name{calc_work_allocation_energy}
\alias{calc_work_allocation_energy}
\title{Calculate work energy requirements (for biophysical allocation)}
\usage{
calc_work_allocation_energy(
  species_short,
  cohort_stock_size = NA_real_,
  metabolic_energy_req_work,
  simulation_duration,
  ratio_me_to_ne = NA_real_
)
}
\arguments{
\item{species_short}{Character. Code identifying the livestock species.
Supported values include:
\itemize{
\item \code{PGS}: pigs
\item \code{CML}: camels
\item \code{CTL}: cattle
\item \code{BFL}: buffalo
\item \code{SHP}: sheep
\item \code{GTS}: goats
}}

\item{cohort_stock_size}{Numeric. Population size in the cohort at the start of the assessment period (heads).}

\item{metabolic_energy_req_work}{Numeric. Energy required for work, used to estimate the energy required for draught
power for CTL, BFL and CML. (MJ/head/day) Assumed to be 0 for other species. Expressed as net energy for CTL, BFL,
SHP, GTS and as metabolizable energy for CML and PGS.}

\item{simulation_duration}{Numeric. Length of the assessment period (days).}

\item{ratio_me_to_ne}{Numeric. Ratio of metabolizable energy converted to net energy (fraction).}
}
\value{
Numeric. Energy required to provide all draught power (traction/work) by cohort (MJ/cohort/assessment
period).
Non-zero values are expected only for draught or work-producing species (CTL, BFL CML).
}
\description{
Calculates the energy required for animal work over the assessment period
(MJ/cohort/assessment period), based on the daily energy requirement for work,
cohort size, and assessment duration.
}
\details{
This function provides the work-related energy term used in a biophysical
allocation framework to apportion emissions between milk and other co-products
in multifunctional livestock production systems.

The approach implements the IDF (2022) standard, adapted from Thoma and Nemecek (2020), and is consistent with
FAO LEAP livestock LCA guidelines (FAO, 2016a, 2016b, 2016c) and with ISO 14044:2006 (Section 4.3.4.2, Step 2).

In accordance with ISO 14044:2006 (Section 4.3.4.2, Step 2), known processing or
biophysical relationships may be used to assign shared inputs and outputs of a
single production unit to individual products or sub-units. In livestock systems,
this includes apportioning shared feed and energy use according to physiological
energy requirements (e.g., net energy for lactation, growth, etc.). If the
resulting process remains multifunctional, these energy terms may subsequently
be used to derive allocation factors among co-products.

Total work-related energy is computed for species (\code{CTL}, \code{BFL}, \code{CML})
and cohorts (, \code{FA}, \code{MA}) assumed to be potentially involved in draught power generation.

The \code{work_allocation_energy} is calculated as follows:

\eqn{energy\_allocation\_work =
energy\_requirement\_work \times simulation\_duration \times cohort\_stock\_size}

for cattle (\code{CTL}) and buffalo (\code{BFL}), and:

\eqn{energy\_allocation\_work =
\frac{energy\_requirement\_work \times simulation\_duration \times cohort\_stock\_size}
{ratio\_me\_to\_ne}}

for camels (\code{CML}).

where \code{metabolic_energy_req_work} can be computed using
\code{\link{calc_metabolic_energy_req_work}} (see also
\code{\link{run_metabolic_energy_req_module}}).

This function is part of the \code{\link[=run_allocation_module]{run_allocation_module()}}.
}
\references{
ISO. (2006). \emph{Environmental management — Life cycle assessment —
Requirements and guidelines (ISO 14044:2006)}. International Organization for
Standardization, Geneva.

IDF. (2022). \emph{The IDF Global Carbon Footprint Standard for the Dairy
Sector}. Bulletin of the IDF No. 520/2022. International Dairy Federation,
Brussels.

Thoma, G., and Nemecek, T. (2020). Allocation between milk and meat in dairy
LCA: Critical discussion of the IDF’s standard methodology. In
\emph{Proceedings of the 12th International Conference on Life Cycle Assessment
of Food (LCAFood 2020)} (pp. 83--89), 13--16 October, Berlin, Germany.

FAO. (2016a). \emph{Environmental performance of large ruminant supply chains:
Guidelines for assessment}. Livestock Environmental Assessment and Performance
(LEAP) Partnership. FAO, Rome, Italy.

FAO. (2016b). \emph{Greenhouse gas emissions and fossil energy use from small
ruminant supply chains: Guidelines for assessment}. Livestock Environmental
Assessment and Performance (LEAP) Partnership. FAO, Rome, Italy.

FAO. (2016c). \emph{Greenhouse gas emissions and fossil energy use from poultry
supply chains: Guidelines for assessment}. Livestock Environmental Assessment
and Performance (LEAP) Partnership. FAO, Rome, Italy.
}
\seealso{
\code{\link{run_allocation_module}},
\code{\link{calc_metabolic_energy_req_work}},
\code{\link{run_metabolic_energy_req_module}}
}
