% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{check_expr_meta}
\alias{check_expr_meta}
\title{Validate expression and metadata alignment}
\usage{
check_expr_meta(expr, meta = NULL)
}
\arguments{
\item{expr}{Expression matrix.}

\item{meta}{Metadata data.frame.}
}
\value{
Invisibly TRUE.
}
\description{
Validate expression and metadata alignment
}
\keyword{internal}
