% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_utils.R
\name{check_geneset}
\alias{check_geneset}
\title{Validate genesets by size}
\usage{
check_geneset(gs, min_genes = 5, max_genes = 500)
}
\arguments{
\item{gs}{Named list genesets.}

\item{min_genes}{Minimum genes per pathway.}

\item{max_genes}{Maximum genes per pathway.}
}
\value{
Filtered named list.
}
\description{
Validate genesets by size
}
