% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{check_input_mode}
\alias{check_input_mode}
\title{Validate input mode}
\usage{
check_input_mode(object, expr, seurat = TRUE)
}
\arguments{
\item{object}{Seurat object.}

\item{expr}{Expression matrix.}

\item{seurat}{Logical, whether Seurat mode is used.}
}
\value{
Invisibly TRUE.
}
\description{
Validate input mode
}
\keyword{internal}
