% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_utils.R
\name{collect_scores}
\alias{collect_scores}
\title{Collect score objects from multiple methods}
\usage{
collect_scores(...)
}
\arguments{
\item{...}{Named \code{gleam_score} objects.}
}
\value{
Named list of score objects.
}
\description{
Collect score objects from multiple methods
}
