% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_compare.R
\name{compare_methods}
\alias{compare_methods}
\title{Alias to compare scoring methods}
\usage{
compare_methods(...)
}
\arguments{
\item{...}{Named score objects from different methods.}
}
\value{
Data.frame.
}
\description{
Alias to compare scoring methods
}
