% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_simulation_core_model.R
\name{compute_fecundity_rates}
\alias{compute_fecundity_rates}
\title{Compute Daily Fecundity Rates}
\usage{
compute_fecundity_rates(parturition_rate, litsize, fem_birth_fraction)
}
\arguments{
\item{parturition_rate}{Numeric. Parturition rate (probability of an adult female to give birth).}

\item{litsize}{Numeric. Prolificacy rate (mean number of offspring per parturition).}

\item{fem_birth_fraction}{Numeric. Female birth ratio (probability that an offspring is female).}
}
\value{
A named list with two elements:
\describe{
\item{fem_fec}{Daily number of female offspring per adult female.}
\item{mal_fec}{Daily number of male offspring per adult female.}
}
}
\description{
Calculates the daily number of male and female offspring produced per adult female.
}
\examples{
compute_fecundity_rates(parturition_rate = 0.8, litsize = 2, fem_birth_fraction = 0.5)

}
