% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{detect_seurat_version}
\alias{detect_seurat_version}
\title{Detect Seurat major version}
\usage{
detect_seurat_version(object = NULL)
}
\arguments{
\item{object}{Seurat object.}
}
\value{
Integer major version (4, 5) when detectable, otherwise \code{NA_integer_}.
}
\description{
Detect Seurat major version
}
\keyword{internal}
