% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_core.R
\name{differential_pathway}
\alias{differential_pathway}
\title{Unified differential pathway interface}
\usage{
differential_pathway(...)
}
\value{
An object of class \code{gleam_test}.
}
\description{
Unified differential pathway interface
}
