% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_spatial.R
\name{differential_pathway_spatial}
\alias{differential_pathway_spatial}
\title{Differential pathway for spatial metadata groups}
\usage{
differential_pathway_spatial(...)
}
\value{
\code{gleam_test} object.
}
\description{
Differential pathway for spatial metadata groups
}
