% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{effect_median_diff}
\alias{effect_median_diff}
\title{Median difference effect size}
\usage{
effect_median_diff(x, g)
}
\arguments{
\item{x}{Numeric score vector.}

\item{g}{Group vector with two groups.}
}
\value{
Numeric effect size.
}
\description{
Median difference effect size
}
\keyword{internal}
