% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{effect_rank_biserial}
\alias{effect_rank_biserial}
\title{Rank-biserial effect size placeholder}
\usage{
effect_rank_biserial(x, g)
}
\arguments{
\item{x}{Numeric vector.}

\item{g}{Group vector.}
}
\value{
Numeric or NA.
}
\description{
Rank-biserial effect size placeholder
}
\keyword{internal}
