% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{effect_standardized_diff}
\alias{effect_standardized_diff}
\title{Standardized mean difference placeholder}
\usage{
effect_standardized_diff(x, g)
}
\arguments{
\item{x}{Numeric vector.}

\item{g}{Group vector.}
}
\value{
Numeric or NA.
}
\description{
Standardized mean difference placeholder
}
\keyword{internal}
