% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_embedding.R
\name{extract_embedding}
\alias{extract_embedding}
\title{Extract embedding matrix}
\usage{
extract_embedding(object, reduction = "umap")
}
\arguments{
\item{object}{Seurat object.}

\item{reduction}{Reduction name.}
}
\value{
Embedding matrix.
}
\description{
Extract embedding matrix
}
