% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_extract.R
\name{extract_embeddings}
\alias{extract_embeddings}
\title{Extract embeddings from generic input}
\usage{
extract_embeddings(
  object = NULL,
  embeddings = NULL,
  reduction = "umap",
  seurat = TRUE
)
}
\arguments{
\item{object}{Seurat object.}

\item{embeddings}{Numeric matrix, optional direct input.}

\item{reduction}{Reduction name when extracting from Seurat.}

\item{seurat}{Input mode.}
}
\value{
Embedding matrix.
}
\description{
Extract embeddings from generic input
}
\keyword{internal}
