% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_extract.R
\name{extract_expr}
\alias{extract_expr}
\title{Extract expression matrix from input}
\usage{
extract_expr(
  object = NULL,
  expr = NULL,
  assay = NULL,
  layer = NULL,
  slot = NULL,
  seurat = TRUE
)
}
\arguments{
\item{object}{Seurat object.}

\item{expr}{Matrix input when \code{seurat = FALSE}.}

\item{assay}{Assay name.}

\item{layer}{Layer name.}

\item{slot}{Legacy slot.}

\item{seurat}{Input mode.}
}
\value{
Matrix or \code{dgCMatrix}.
}
\description{
Extract expression matrix from input
}
\keyword{internal}
