% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_utils.R
\name{extract_lineage}
\alias{extract_lineage}
\title{Extract lineage labels}
\usage{
extract_lineage(score, lineage = NULL, object = NULL)
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{lineage}{Metadata column name, vector, or NULL.}

\item{object}{Optional trajectory object.}
}
\value{
Character vector.
}
\description{
Extract lineage labels
}
