% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_extract.R
\name{extract_meta}
\alias{extract_meta}
\title{Extract metadata from input}
\usage{
extract_meta(object = NULL, meta = NULL, expr = NULL, seurat = TRUE)
}
\arguments{
\item{object}{Seurat object.}

\item{meta}{Metadata data.frame for matrix mode.}

\item{expr}{Expression matrix for matrix mode.}

\item{seurat}{Input mode.}
}
\value{
Metadata data.frame aligned to cells.
}
\description{
Extract metadata from input
}
\keyword{internal}
