% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_extract.R
\name{extract_reduction}
\alias{extract_reduction}
\title{Extract reduction embedding matrix}
\usage{
extract_reduction(object, reduction = "umap")
}
\arguments{
\item{object}{Seurat object.}

\item{reduction}{Reduction name (for example: \code{umap}, \code{pca}).}
}
\value{
Numeric embedding matrix.
}
\description{
Extract reduction embedding matrix
}
\keyword{internal}
