% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_extract.R
\name{extract_spatial_coords}
\alias{extract_spatial_coords}
\title{Extract spatial coordinates}
\usage{
extract_spatial_coords(
  object = NULL,
  meta = NULL,
  coords = NULL,
  seurat = TRUE
)
}
\arguments{
\item{object}{Seurat object.}

\item{meta}{Metadata for matrix mode.}

\item{coords}{Optional direct coordinate data.frame.}

\item{seurat}{Input mode.}
}
\value{
Data.frame with columns \code{x} and \code{y}.
}
\description{
Extract spatial coordinates
}
\keyword{internal}
