% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_extract.R
\name{extract_spatial_meta}
\alias{extract_spatial_meta}
\title{Extract spatial metadata}
\usage{
extract_spatial_meta(object = NULL, meta = NULL, seurat = TRUE)
}
\arguments{
\item{object}{Seurat object.}

\item{meta}{Metadata for matrix mode.}

\item{seurat}{Input mode.}
}
\value{
Metadata data.frame.
}
\description{
Extract spatial metadata
}
\keyword{internal}
