% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_msigdb.R
\name{get_geneset_msigdb}
\alias{get_geneset_msigdb}
\title{Get MSigDB genesets via msigdbr}
\usage{
get_geneset_msigdb(
  species = "Homo sapiens",
  collection = "H",
  subcollection = NULL
)
}
\arguments{
\item{species}{Species label accepted by \code{msigdbr}.}

\item{collection}{MSigDB collection, e.g. \code{H}, \code{C2}, \code{C5}.}

\item{subcollection}{Optional subcollection, e.g. \code{CP:KEGG}.}
}
\value{
Named list genesets.
}
\description{
Get MSigDB genesets via msigdbr
}
\keyword{internal}
