% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_utils.R
\name{get_palette}
\alias{get_palette}
\title{Get palette colors}
\usage{
get_palette(
  name = "gleam_discrete",
  n = 8,
  continuous = FALSE,
  reverse = FALSE
)
}
\arguments{
\item{name}{Palette name.}

\item{n}{Number of colors.}

\item{continuous}{Whether to return continuous palette function output.}

\item{reverse}{Reverse color order.}
}
\value{
Character color vector.
}
\description{
Get palette colors
}
