% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleam-package.R
\docType{package}
\name{gleam-package}
\alias{gleam}
\alias{gleam-package}
\title{gleam: Global Livestock Environmental Assessment Model (GLEAM-X)}
\description{
GLEAM-X is the official R implementation of FAO’s Global Livestock Environmental Assessment Model (GLEAM). The package provides a modular, transparent framework for simulating livestock production systems and quantifying their environmental impacts. Methodological background: MacLeod et al. (2017) "Invited review: A position on the Global Livestock Environmental Assessment Model (GLEAM)" \doi{10.1017/S1751731117001847}. Further information: \url{https://www.fao.org/gleam/en/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/un-fao/GLEAM/}
  \item \url{https://www.fao.org/gleam/en/}
  \item Report bugs at \url{https://github.com/un-fao/GLEAM/issues}
}

}
\author{
\strong{Maintainer}: Ahmed Jou \email{ahmed@applitics.fr} (\href{https://orcid.org/0009-0002-2309-4583}{ORCID}) (https://applitics.fr)

Authors:
\itemize{
  \item Ahmed Jou \email{ahmed@applitics.fr} (\href{https://orcid.org/0009-0002-2309-4583}{ORCID}) (https://applitics.fr)
  \item Yassine Elaouni \email{yassine@applitics.fr} (https://applitics.fr)
  \item Dominik Wisser \email{dominik.wisser@fao.org} (\href{https://orcid.org/0000-0001-8368-3801}{ORCID}) (GLEAM team lead, FAO)
  \item Lydia Lanzoni \email{lydia.lanzoni@fao.org} (\href{https://orcid.org/0000-0002-4488-4368}{ORCID}) (Livestock and climate change specialist, FAO)
  \item Giuseppe Tempio \email{giuseppe.tempio@fao.org} (\href{https://orcid.org/0009-0004-2272-0794}{ORCID}) (Livestock environmental impact analyst, FAO)
}

Other contributors:
\itemize{
  \item Phyllis Ndungu' \email{Phyllis.Ndungu@fao.org} (Livestock and climate change specialist, FAO) [contributor]
  \item Yushan Li \email{Yushan.Li@fao.org} (Livestock and sustainability specialist, FAO) [contributor]
  \item Manling Xu \email{Manling.Xu@fao.org} (Livestock data intern, FAO) [contributor]
  \item Zixin Wang \email{Zixin.Wang@fao.org} (Livestock data intern, FAO) [contributor]
  \item Narindra Rakotovao \email{Narindra.Rakotovao@fao.org} (Livestock policy officer, FAO) [contributor]
  \item Stewart Chikoloma \email{Stewart.Chikomola@fao.org} (Livestock data specialist, FAO) [contributor]
  \item David OBrien \email{david@applitics.fr} (\href{https://orcid.org/0009-0007-0339-0517}{ORCID}) (https://applitics.fr) [contributor]
  \item Food and Agriculture Organization of the United Nations FAO (Original GLEAM model framework) [copyright holder]
  \item Federal Ministry of Agriculture, Food and Regional Identity of Germany BMELH (Funding support) [funder]
}

}
\keyword{internal}
