% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleam_constants.R
\name{gleam_cohorts}
\alias{gleam_cohorts}
\title{Valid cohort short codes}
\format{
A character vector of length 8.
}
\usage{
gleam_cohorts
}
\description{
All sex- and age-class cohort codes recognised by GLEAM. Derived from
\code{gleam_cohorts_names}.
}
\keyword{internal}
