% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleam_constants.R
\name{gleam_cohorts_demographic}
\alias{gleam_cohorts_demographic}
\title{Demographic cohort short codes}
\format{
A character vector of length 6.
}
\usage{
gleam_cohorts_demographic
}
\description{
The 6 sex- and age-class cohort codes used by the demographic herd module.
}
\keyword{internal}
