% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleam_constants.R
\name{gleam_species}
\alias{gleam_species}
\title{Valid species short codes}
\format{
A character vector of length 6.
}
\usage{
gleam_species
}
\description{
All livestock species supported by GLEAM. Derived from \code{gleam_species_names}.
}
\keyword{internal}
