% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleam_constants.R
\name{gleam_species_milk_producers}
\alias{gleam_species_milk_producers}
\title{Milk-producing species}
\format{
A character vector of length 5.
}
\usage{
gleam_species_milk_producers
}
\description{
Species for which milk production, lactation energy requirements, and
ruminant-style digestibility parameters apply. Includes ruminants plus
Camels.
}
\keyword{internal}
