% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleam_constants.R
\docType{data}
\name{gleam_species_poultry}
\alias{gleam_species_poultry}
\title{Poultry species}
\format{
A character vector of length 1.
}
\usage{
gleam_species_poultry
}
\description{
Species using the poultry-specific demographic energy and nitrogen equations.
}
\keyword{internal}
