% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hallmark}
\alias{hallmark}
\title{Built-in hallmark-like gene sets}
\format{
A named list.
}
\source{
Internal toy collection
}
\usage{
hallmark
}
\description{
A named list of hallmark-like pathways and genes.
}
\keyword{datasets}
