% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immune_small}
\alias{immune_small}
\title{Built-in immune-focused small gene sets}
\format{
A named list.
}
\source{
Internal toy collection
}
\usage{
immune_small
}
\description{
A named list of immune pathways and genes.
}
\keyword{datasets}
