% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{is_seurat_object}
\alias{is_seurat_object}
\title{Check whether object is Seurat-like}
\usage{
is_seurat_object(object)
}
\arguments{
\item{object}{Input object.}
}
\value{
Logical.
}
\description{
Check whether object is Seurat-like
}
\keyword{internal}
