% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_extract.R
\name{is_spatial_object}
\alias{is_spatial_object}
\title{Check whether input contains spatial information}
\usage{
is_spatial_object(object = NULL, meta = NULL)
}
\arguments{
\item{object}{Seurat object or list.}

\item{meta}{Metadata data.frame.}
}
\value{
Logical.
}
\description{
Check whether input contains spatial information
}
\keyword{internal}
