% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_utils.R
\name{join_score_trajectory}
\alias{join_score_trajectory}
\title{Join score matrix and trajectory metadata}
\usage{
join_score_trajectory(score, pseudotime = NULL, lineage = NULL)
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{pseudotime}{Pseudotime source.}

\item{lineage}{Lineage source.}
}
\value{
Long-format data.frame.
}
\description{
Join score matrix and trajectory metadata
}
