% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_sources.R
\name{list_geneset_sources}
\alias{list_geneset_sources}
\title{List geneset sources}
\usage{
list_geneset_sources()
}
\value{
Data.frame of supported geneset sources.
}
\description{
List geneset sources
}
