% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_registry.R
\name{list_scoring_methods}
\alias{list_scoring_methods}
\title{List available scoring methods}
\usage{
list_scoring_methods(include_optional = TRUE)
}
\arguments{
\item{include_optional}{Whether to include optional integration methods.}
}
\value{
A data.frame describing methods and dependencies.
}
\description{
List available scoring methods
}
