% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_utils.R
\name{map_scores_to_trajectory}
\alias{map_scores_to_trajectory}
\title{Map scores to trajectory table}
\usage{
map_scores_to_trajectory(score, pseudotime = NULL, lineage = NULL)
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{pseudotime}{Pseudotime source.}

\item{lineage}{Lineage source.}
}
\value{
Long-format data.frame.
}
\description{
Map scores to trajectory table
}
