% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_utils.R
\name{match_geneset}
\alias{match_geneset}
\title{Match genesets to expression genes}
\usage{
match_geneset(gs, expr_genes, verbose = TRUE)
}
\arguments{
\item{gs}{Named list genesets.}

\item{expr_genes}{Character vector of expression rownames.}

\item{verbose}{Print overlap summary.}
}
\value{
List with matched genesets and overlap info.
}
\description{
Match genesets to expression genes
}
