% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_score.R
\name{new_scpathway_score}
\alias{new_scpathway_score}
\title{Construct score object}
\usage{
new_scpathway_score(
  score,
  meta,
  method,
  geneset_name,
  geneset_info,
  params = list()
)
}
\arguments{
\item{score}{Pathway-by-cell matrix.}

\item{meta}{Cell metadata \code{data.frame} aligned to score columns.}

\item{method}{Scoring method.}

\item{geneset_name}{Geneset label.}

\item{geneset_info}{Geneset metadata.}

\item{params}{Parameter list.}
}
\value{
Object of class \code{gleam_score} and \code{gleam_score}.
}
\description{
Construct score object
}
\keyword{internal}
