% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_test.R
\name{new_scpathway_test}
\alias{new_scpathway_test}
\title{Construct differential test object}
\usage{
new_scpathway_test(table, level, method, comparison = list(), params = list())
}
\arguments{
\item{table}{Differential pathway result table.}

\item{level}{Comparison level.}

\item{method}{Statistical method.}

\item{comparison}{Comparison details.}

\item{params}{Parameter list.}
}
\value{
Object of class \code{gleam_test} and \code{gleam_test}.
}
\description{
Construct differential test object
}
\keyword{internal}
