% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_medium_matrix}
\alias{pbmc_medium_matrix}
\title{Medium PBMC expression matrix for realistic demos}
\format{
A \code{dgCMatrix}.
}
\source{
Simulated PBMC-like data
}
\usage{
pbmc_medium_matrix
}
\description{
Sparse gene-by-cell matrix suitable for tutorial workflows.
}
\keyword{datasets}
