% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_medium_meta}
\alias{pbmc_medium_meta}
\title{Medium PBMC metadata for realistic demos}
\format{
A \code{data.frame}.
}
\source{
Simulated PBMC-like data
}
\usage{
pbmc_medium_meta
}
\description{
Metadata aligned to \code{pbmc_medium_matrix} columns.
}
\keyword{datasets}
