% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_utils.R
\name{pivot_scores_long}
\alias{pivot_scores_long}
\title{Pivot score matrix to long format}
\usage{
pivot_scores_long(score)
}
\arguments{
\item{score}{\code{gleam_score} object.}
}
\value{
Long-format data.frame with \code{cell_id}, \code{pathway}, and \code{score}.
}
\description{
Pivot score matrix to long format
}
