% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plot_box}
\alias{plot_box}
\title{Box plot of sample-level pathway scores}
\usage{
plot_box(score, pathway, group, sample, palette = "gleam_discrete")
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{pathway}{Pathway name.}

\item{group}{Group variable (metadata column name or vector).}

\item{sample}{Sample variable (metadata column name or vector).}

\item{palette}{Discrete palette name or custom colors.}
}
\value{
A \code{ggplot} object.
}
\description{
Box plot of sample-level pathway scores
}
