% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_celltype_compare.R
\name{plot_celltype_compare}
\alias{plot_celltype_compare}
\title{Plot celltype comparison result}
\usage{
plot_celltype_compare(x, top_n = 20)
}
\arguments{
\item{x}{\code{gleam_test} object or result table.}

\item{top_n}{Number of pathways to display.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot celltype comparison result
}
