% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot_bar.R
\name{plot_dot_bar}
\alias{plot_dot_bar}
\title{Dot + bar summary plot for pathway groups}
\usage{
plot_dot_bar(score, by, threshold = 0, pathway = NULL)
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{by}{Grouping metadata columns.}

\item{threshold}{Threshold for active fraction.}

\item{pathway}{Optional pathway filter.}
}
\value{
A \code{ggplot} object.
}
\description{
Dot + bar summary plot for pathway groups
}
