% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedding_score.R
\name{plot_embedding_score}
\alias{plot_embedding_score}
\title{Plot pathway score on embedding coordinates}
\usage{
plot_embedding_score(score, pathway, embedding, split.by = NULL)
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{pathway}{Pathway name.}

\item{embedding}{Embedding matrix with at least 2 columns.}

\item{split.by}{Optional metadata column for faceting.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot pathway score on embedding coordinates
}
