% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_in_celltype.R
\name{plot_group_in_celltype}
\alias{plot_group_in_celltype}
\title{Plot group comparison within a fixed celltype}
\usage{
plot_group_in_celltype(x, top_n = 20, alpha = 0.85, theme_params = list())
}
\arguments{
\item{x}{\code{gleam_test} object or result table.}

\item{top_n}{Number of signatures to display.}

\item{alpha}{Bar alpha.}

\item{theme_params}{Optional list passed to \code{\link[=gleam_theme]{gleam_theme()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot group comparison within a fixed celltype
}
