% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pseudobulk_box.R
\name{plot_pseudobulk_box}
\alias{plot_pseudobulk_box}
\title{Pseudobulk boxplot for pathway scores}
\usage{
plot_pseudobulk_box(score, pathway, group, sample, celltype = NULL)
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{pathway}{Pathway name.}

\item{group}{Group variable.}

\item{sample}{Sample variable.}

\item{celltype}{Optional celltype variable.}
}
\value{
A \code{ggplot} object.
}
\description{
Pseudobulk boxplot for pathway scores
}
