% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pseudotime_score.R
\name{plot_pseudotime_score}
\alias{plot_pseudotime_score}
\title{Plot pathway score over pseudotime}
\usage{
plot_pseudotime_score(
  score,
  pathway,
  pseudotime = NULL,
  lineage = NULL,
  smooth = TRUE
)
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{pathway}{Pathway name.}

\item{pseudotime}{Pseudotime source.}

\item{lineage}{Optional lineage source for coloring.}

\item{smooth}{Add smoothing line.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot pathway score over pseudotime
}
