% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial_multi.R
\name{plot_spatial_multi}
\alias{plot_spatial_multi}
\title{Plot multiple pathways on spatial coordinates}
\usage{
plot_spatial_multi(score, pathways, coords, palette = "gleam_continuous")
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{pathways}{Character vector of pathway names.}

\item{coords}{Spatial coordinates data.frame with x/y.}

\item{palette}{Continuous palette.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot multiple pathways on spatial coordinates
}
