% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial_score.R
\name{plot_spatial_score}
\alias{plot_spatial_score}
\title{Plot pathway score on spatial coordinates}
\usage{
plot_spatial_score(
  score,
  pathway,
  coords,
  image = NULL,
  palette = "gleam_continuous",
  split.by = NULL
)
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{pathway}{Pathway name.}

\item{coords}{Spatial coordinates data.frame with x/y.}

\item{image}{Optional raster/image object. If provided, used as background.}

\item{palette}{Continuous palette name or colors.}

\item{split.by}{Optional metadata split variable.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot pathway score on spatial coordinates
}
