% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_split_violin.R
\name{plot_split_violin}
\alias{plot_split_violin}
\title{Split violin plot for pathway scores}
\usage{
plot_split_violin(score, pathway, x, split.by)
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{pathway}{Pathway name.}

\item{x}{Grouping variable for x-axis.}

\item{split.by}{Variable used for split/fill.}
}
\value{
A \code{ggplot} object.
}
\description{
Split violin plot for pathway scores
}
