% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_simulation_core_model.R
\name{project_population_size}
\alias{project_population_size}
\title{Project One Year of Steady-State Population Dynamics}
\usage{
project_population_size(
  herd_size_total,
  fecundity_female,
  fecundity_male,
  probability_death,
  probability_offtake,
  probability_growth,
  growth_rate_herd,
  herd_structure,
  cohort_share
)
}
\arguments{
\item{herd_size_total}{Numeric. Total population size at the start of the year.}

\item{fecundity_female}{Numeric. Daily female births per adult female.}

\item{fecundity_male}{Numeric. Daily male births per adult female.}

\item{probability_death}{Named numeric vector of length 10. Daily death probabilities. Must be named using:
\code{FB}, \code{FJ}, \code{FS}, \code{FA}, \code{FC}, \code{MB}, \code{MJ}, \code{MS}, \code{MA}, \code{MC}.}

\item{probability_offtake}{Named numeric vector of length 10. Daily offtake probabilities.
Names must match those in \code{probability_death}.}

\item{probability_growth}{Named numeric vector of length 10. Transition probabilities to next age class.
Names must match those in \code{probability_death}.}

\item{growth_rate_herd}{Numeric. Annual population growth rate.}

\item{herd_structure}{Named numeric vector of length 8. Final population share in 8 classes. Must be named:
\code{FB}, \code{FJ}, \code{FS}, \code{FA}, \code{MB}, \code{MJ}, \code{MS}, \code{MA}.}

\item{cohort_share}{Named numeric vector of length 6. Final population share in 6 grouped classes. Must be named:
\code{FJ}, \code{FS}, \code{FA}, \code{MJ}, \code{MS}, \code{MA}.}
}
\value{
A named list with:
\describe{
\item{cohort_stock_start}{Size in 6 sex-age classes at the start of the year.}
\item{cohort_stock_end_projected}{Size at year end (projected using growth rate).}
\item{cohort_stock_end_exact_simulated}{Size at year end (based on full simulation over 366 days).}
\item{cohort_stock_average}{Average population size over the year.}
\item{cohort_offtake_heads}{Total offtake over the year per sex-age class.}
}
}
\description{
Simulates one year of population dynamics under steady-state assumptions using demographic parameters
and returns population size statistics and offtake results.
}
