% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_indirectemissions_feed.R
\name{run_feed_production_emissions}
\alias{run_feed_production_emissions}
\title{Calculate Feed Production GHG Emissions}
\usage{
run_feed_production_emissions(rations_share, feed_emissions)
}
\arguments{
\item{rations_share}{A data.table containing feed shares per cohort. Must include:
\itemize{
\item \code{herd_id}, \code{animal}, \code{feed_id}, \code{cohort}, \code{feed_ration_fraction}.
}}

\item{feed_emissions}{A data.table of feed production emission factors. Must include:
\itemize{
\item \code{feed_id},
\item \code{co2_feed_fertilizer}, \code{co2_feed_pesticides},
\item \code{co2_feed_crop_operations}, \code{co2_feed_luc_nopeat}, \code{co2_feed_luc_peat},
\item \code{n2o_feed_fertilizer}, \code{n2o_feed_manure_applied}, \code{n2o_feed_crop_residues},
\item \code{ch4_feed_rice}.
}}
}
\value{
A data.table summarized by \code{herd_id}, \code{animal}, and \code{cohort} with:
\itemize{
\item \code{diet_co2_feed_fertilizer}, \code{diet_co2_feed_pesticides},
\item \code{diet_co2_feed_crop_operations}, \code{diet_co2_feed_luc_nopeat},
\item \code{diet_co2_feed_luc_peat},
\item \code{diet_n2o_feed_fertilizer}, \code{diet_n2o_feed_manure_applied},
\item \code{diet_n2o_feed_crop_residues},
\item \code{diet_ch4_feed_rice}
}
}
\description{
Computes cohort-level CO2, N2O, and CH4 emissions from feed production
(fertilizer, pesticides, crop operations, land-use change, manure, residues, rice)
based on feed rations and feed emission factors.
}
\examples{
\dontrun{
# Load cleaned example input from the package and compute feed intake metrics#' 
feed_emissions <- data.table::fread(
  system.file("extdata/Parameters/feed/feed_emission_factors.csv", package = "gleam")
)


rations_share <- data.table::fread(
  system.file("extdata/examples/feed_rations_share_example.csv", package = "gleam")
)

result <- run_feed_production_emissions(rations_share, feed_emissions)
}
}
