% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_utils.R
\name{scale_gleam_color}
\alias{scale_gleam_color}
\title{GLEAM color scale helper}
\usage{
scale_gleam_color(palette = "gleam_discrete", continuous = FALSE)
}
\arguments{
\item{palette}{Palette name or custom color vector.}

\item{continuous}{Whether to return continuous scale.}
}
\value{
A ggplot2 scale object.
}
\description{
GLEAM color scale helper
}
