% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_utils.R
\name{scale_gleam_fill}
\alias{scale_gleam_fill}
\title{GLEAM fill scale helper}
\usage{
scale_gleam_fill(palette = "gleam_discrete", continuous = FALSE)
}
\arguments{
\item{palette}{Palette name or custom color vector.}

\item{continuous}{Whether to return continuous scale.}
}
\value{
A ggplot2 scale object.
}
\description{
GLEAM fill scale helper
}
