% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_auc.R
\name{score_auc_matrix}
\alias{score_auc_matrix}
\title{AUC-like pathway scoring on matrix input}
\usage{
score_auc_matrix(expr, genesets, auc_max_rank = 0.05, verbose = TRUE)
}
\arguments{
\item{expr}{Gene-by-cell matrix.}

\item{genesets}{Named list of pathways.}

\item{auc_max_rank}{Fraction of top-ranked genes considered active.}

\item{verbose}{Whether to print progress.}
}
\value{
Pathway-by-cell numeric matrix.
}
\description{
AUC-like pathway scoring on matrix input
}
\keyword{internal}
