% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_mean.R
\name{score_mean_matrix}
\alias{score_mean_matrix}
\title{Mean pathway scoring on matrix input}
\usage{
score_mean_matrix(expr, genesets, verbose = TRUE)
}
\arguments{
\item{expr}{Gene-by-cell matrix.}

\item{genesets}{Named list of pathways.}

\item{verbose}{Whether to print progress.}
}
\value{
Pathway-by-cell numeric matrix.
}
\description{
Mean pathway scoring on matrix input
}
\keyword{internal}
