% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_rank.R
\name{score_rank_matrix}
\alias{score_rank_matrix}
\title{Rank-based pathway scoring on matrix input}
\usage{
score_rank_matrix(expr, genesets, normalize = TRUE, verbose = TRUE)
}
\arguments{
\item{expr}{Gene-by-cell matrix.}

\item{genesets}{Named list of pathways.}

\item{normalize}{Whether to normalize rank score to the 0-1 range.}

\item{verbose}{Whether to print progress.}
}
\value{
Pathway-by-cell numeric matrix.
}
\description{
Rank-based pathway scoring on matrix input
}
\keyword{internal}
