% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_scaled_mean.R
\name{score_robust_matrix}
\alias{score_robust_matrix}
\title{Robust normalized pathway scoring}
\usage{
score_robust_matrix(expr, genesets, verbose = TRUE)
}
\arguments{
\item{expr}{Gene-by-cell matrix.}

\item{genesets}{Named list of pathways.}

\item{verbose}{Whether to print progress.}
}
\value{
Pathway-by-cell numeric matrix.
}
\description{
Robust normalized pathway scoring
}
\keyword{internal}
