% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_singscore.R
\name{score_singscore_like_matrix}
\alias{score_singscore_like_matrix}
\title{Singscore-like pathway scoring on matrix input}
\usage{
score_singscore_like_matrix(expr, genesets, verbose = TRUE)
}
\arguments{
\item{expr}{Gene-by-cell matrix.}

\item{genesets}{Named list of pathways.}

\item{verbose}{Whether to print progress.}
}
\value{
Pathway-by-cell numeric matrix.
}
\description{
Singscore-like pathway scoring on matrix input
}
\keyword{internal}
