% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_zmean.R
\name{score_zmean_matrix}
\alias{score_zmean_matrix}
\title{Z-mean pathway scoring on matrix input}
\usage{
score_zmean_matrix(expr, genesets, verbose = TRUE)
}
\arguments{
\item{expr}{Gene-by-cell matrix.}

\item{genesets}{Named list of pathways.}

\item{verbose}{Whether to print progress.}
}
\value{
Pathway-by-cell numeric matrix.
}
\description{
Z-mean pathway scoring on matrix input
}
\keyword{internal}
