% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_zscore.R
\name{score_zscore_matrix}
\alias{score_zscore_matrix}
\title{Z-score pathway scoring on matrix input}
\usage{
score_zscore_matrix(expr, genesets, verbose = TRUE)
}
\arguments{
\item{expr}{Gene-by-cell matrix.}

\item{genesets}{Named list of pathways.}

\item{verbose}{Whether to print progress.}
}
\value{
Pathway-by-cell numeric matrix.
}
\description{
Z-score pathway scoring on matrix input
}
\keyword{internal}
