% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPathway-package.R
\docType{package}
\name{scPathway-package}
\alias{scPathway}
\alias{scPathway-package}
\title{scPathway: Lightweight pathway analysis for single-cell RNA-seq}
\description{
Matrix-native pathway scoring and comparison toolkit with optional Seurat v5
compatibility support.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jameswoo/scPathway}
  \item Report bugs at \url{https://github.com/jameswoo/scPathway/issues}
}

}
\author{
\strong{Maintainer}: James Woo \email{james@example.com}

}
\keyword{internal}
