% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_sources.R
\name{search_geneset}
\alias{search_geneset}
\title{Search pathways in a geneset collection}
\usage{
search_geneset(geneset = "hallmark", pattern, ignore.case = TRUE, ...)
}
\arguments{
\item{geneset}{Geneset input accepted by \code{\link[=get_geneset]{get_geneset()}}.}

\item{pattern}{Search pattern.}

\item{ignore.case}{Whether search is case-insensitive.}

\item{...}{Additional arguments passed to \code{\link[=get_geneset]{get_geneset()}}.}
}
\value{
Character vector of matched pathway names.
}
\description{
Search pathways in a geneset collection
}
