% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_seurat_v4_v5.R
\name{seurat_mode}
\alias{seurat_mode}
\title{Check Seurat compatibility mode}
\usage{
seurat_mode(object)
}
\arguments{
\item{object}{Seurat object.}
}
\value{
Character string describing detected Seurat mode.
}
\description{
Check Seurat compatibility mode
}
