% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_medium_coords}
\alias{spatial_medium_coords}
\title{Medium spatial coordinates}
\format{
A \code{data.frame} with columns \code{x} and \code{y}.
}
\source{
Simulated spatial transcriptomics-like data
}
\usage{
spatial_medium_coords
}
\description{
Coordinates aligned to spatial spots.
}
\keyword{datasets}
