% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_medium_expr}
\alias{spatial_medium_expr}
\title{Medium spatial expression matrix}
\format{
A \code{dgCMatrix}.
}
\source{
Simulated spatial transcriptomics-like data
}
\usage{
spatial_medium_expr
}
\description{
Sparse gene-by-spot matrix for spatial workflow demos.
}
\keyword{datasets}
