% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_medium_meta}
\alias{spatial_medium_meta}
\title{Medium spatial metadata}
\format{
A \code{data.frame}.
}
\source{
Simulated spatial transcriptomics-like data
}
\usage{
spatial_medium_meta
}
\description{
Metadata for spatial spots including sample/region/domain and coordinates.
}
\keyword{datasets}
