% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_spatial.R
\name{spatial_table}
\alias{spatial_table}
\title{Build spatial table from score object}
\usage{
spatial_table(score, coords, meta = NULL)
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{coords}{Coordinate data.frame with x/y.}

\item{meta}{Optional metadata table.}
}
\value{
Data.frame.
}
\description{
Build spatial table from score object
}
