% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_utils.R
\name{summarize_scores}
\alias{summarize_scores}
\title{Summarize score table}
\usage{
summarize_scores(
  score,
  by = c("sample", "group"),
  fun = c("mean", "median", "fraction"),
  threshold = 0
)
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{by}{Grouping columns.}

\item{fun}{Aggregation function.}

\item{threshold}{Threshold used for fraction.}
}
\value{
Summarized data.frame.
}
\description{
Summarize score table
}
