% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_within_celltype.R
\name{test_groups_within_celltype}
\alias{test_groups_within_celltype}
\title{Within-celltype group comparison}
\usage{
test_groups_within_celltype(
  score_mat,
  group,
  celltype,
  target_celltype,
  sample = NULL,
  level = c("cell", "sample"),
  method = c("wilcox", "t"),
  ref_group = NULL,
  adjust_method = "BH"
)
}
\arguments{
\item{score_mat}{Pathway-by-cell matrix.}

\item{group}{Group vector.}

\item{celltype}{Celltype vector.}

\item{target_celltype}{Target celltype.}

\item{sample}{Optional sample vector.}

\item{level}{Comparison level (\code{cell} or \code{sample}).}

\item{method}{Test method.}

\item{ref_group}{Reference group.}

\item{adjust_method}{P-value adjustment method.}
}
\value{
Differential pathway result data.frame.
}
\description{
Within-celltype group comparison
}
\keyword{internal}
