% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_celltype.R
\name{test_pathway_celltype}
\alias{test_pathway_celltype}
\title{Celltype-level pathway test}
\usage{
test_pathway_celltype(
  score_mat,
  celltype,
  method = c("wilcox", "t"),
  ref_celltype = NULL,
  adjust_method = "BH"
)
}
\arguments{
\item{score_mat}{Pathway-by-cell matrix.}

\item{celltype}{Celltype vector.}

\item{method}{Test method.}

\item{ref_celltype}{Reference celltype.}

\item{adjust_method}{P-value adjustment method.}
}
\value{
Differential pathway result data.frame.
}
\description{
Celltype-level pathway test
}
\keyword{internal}
