% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_model.R
\name{test_pathway_lm_placeholder}
\alias{test_pathway_lm_placeholder}
\title{Experimental linear model placeholder}
\usage{
test_pathway_lm_placeholder(score_mat, formula, data)
}
\arguments{
\item{score_mat}{Pathway-by-cell matrix.}

\item{formula}{Formula for model.}

\item{data}{Data frame.}
}
\value{
A placeholder list.
}
\description{
Experimental linear model placeholder
}
\keyword{internal}
