% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sample.R
\name{test_pathway_sample}
\alias{test_pathway_sample}
\title{Sample-level pathway test}
\usage{
test_pathway_sample(
  score_mat,
  group,
  sample,
  method = c("wilcox", "t"),
  ref_group = NULL,
  adjust_method = "BH"
)
}
\arguments{
\item{score_mat}{Pathway-by-cell matrix.}

\item{group}{Group vector.}

\item{sample}{Sample vector.}

\item{method}{Test method.}

\item{ref_group}{Reference group.}

\item{adjust_method}{P-value adjustment method.}
}
\value{
Differential pathway result data.frame.
}
\description{
Sample-level pathway test
}
\keyword{internal}
