% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_expr}
\alias{toy_expr}
\title{Tiny toy expression dataset for tests}
\format{
A list with components \code{expr} and \code{meta}.
}
\source{
Simulated data
}
\usage{
toy_expr
}
\description{
A list with expression matrix and metadata for fast unit tests.
}
\keyword{datasets}
