% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_trajectory.R
\name{trajectory_table}
\alias{trajectory_table}
\title{Build trajectory table from score object}
\usage{
trajectory_table(score, pseudotime = NULL, lineage = NULL, embeddings = NULL)
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{pseudotime}{Pseudotime source.}

\item{lineage}{Lineage source.}

\item{embeddings}{Optional embedding matrix.}
}
\value{
Data.frame.
}
\description{
Build trajectory table from score object
}
