% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_herd_simulation_core_model.R
\name{validate_avg_weight_inputs}
\alias{validate_avg_weight_inputs}
\title{Validate inputs for calc_avg_weights}
\usage{
validate_avg_weight_inputs(
  initial_weight,
  potential_final_weight,
  slaughter_weight,
  offtake_rate
)
}
\description{
Ensures all arguments are numeric scalars (length 1), allows NA.
}
\keyword{internal}
