% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_herd_simulation_core_model.R
\name{validate_cohort_weight_inputs}
\alias{validate_cohort_weight_inputs}
\title{Validate inputs for calc_cohort_weights}
\usage{
validate_cohort_weight_inputs(
  animal,
  cohort,
  adult_fem_weight,
  adult_mal_weight,
  birth_weight,
  slaughter_weight_fem,
  slaughter_weight_mal,
  weaning_weight,
  age_first_calving,
  animal_age
)
}
\description{
Validate inputs for calc_cohort_weights
}
\keyword{internal}
