% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_herd_simulation_core_model.R
\name{validate_daily_gain_inputs}
\alias{validate_daily_gain_inputs}
\title{Validate inputs for calc_daily_weight_gain}
\usage{
validate_daily_gain_inputs(potential_final_weight, initial_weight, duration)
}
\description{
Ensures arguments are numeric scalars (length 1), allows NA values.
}
\keyword{internal}
