% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_herd_simulation_core_model.R
\name{validate_fecundity_inputs}
\alias{validate_fecundity_inputs}
\title{Validate inputs for compute_fecundity_rates}
\usage{
validate_fecundity_inputs(part_rate, prolif_rate, fem_birth_ratio)
}
\description{
Validate inputs for compute_fecundity_rates
}
\keyword{internal}
