% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_herd_simulation_core_model.R
\name{validate_offtake_summary_inputs}
\alias{validate_offtake_summary_inputs}
\title{Validate inputs for summarise_offtake}
\usage{
validate_offtake_summary_inputs(size, size_end, size_avg, offtake)
}
\description{
Validate inputs for summarise_offtake
}
\keyword{internal}
