% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_herd_simulation_core_model.R
\name{validate_population_size_inputs}
\alias{validate_population_size_inputs}
\title{Validate inputs for project_population_size}
\usage{
validate_population_size_inputs(
  size_total,
  fem_fec,
  mal_fec,
  prob_death,
  prob_offtake,
  prob_growth,
  growth_rate_pop,
  structure,
  share
)
}
\description{
Validate inputs for project_population_size
}
\keyword{internal}
