% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_herd_simulation_core_model.R
\name{validate_scalar_character}
\alias{validate_scalar_character}
\title{Validate a scalar character input}
\usage{
validate_scalar_character(x, arg_name)
}
\arguments{
\item{x}{The object to validate.}

\item{arg_name}{A string. The name of the argument (used in the error message).}
}
\description{
Ensures that the input is a single, non-missing character value.
This function is typically used to validate identifiers or categorical inputs
such as \code{animal} or \code{cohort} within model functions.
}
\keyword{internal}
