% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_herd_simulation_core_model.R
\name{validate_steady_state_inputs}
\alias{validate_steady_state_inputs}
\title{Validate inputs for simulate_steady_state_structure}
\usage{
validate_steady_state_inputs(
  initial_structure,
  max_years,
  min_lambda_change,
  fem_fec,
  mal_fec,
  prob_death,
  prob_offtake,
  prob_growth
)
}
\description{
Validate inputs for simulate_steady_state_structure
}
\keyword{internal}
