% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_herd_simulation_core_model.R
\name{validate_transition_inputs}
\alias{validate_transition_inputs}
\title{Validate inputs for compute_transition_probabilities}
\usage{
validate_transition_inputs(duration, offtake_rate, death_rate)
}
\description{
Validate inputs for compute_transition_probabilities
}
\keyword{internal}
