% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation_core_model.R
\name{aggregate_cohort_to_herd}
\alias{aggregate_cohort_to_herd}
\title{Aggregate Cohort-Level Data to Herd-Level}
\usage{
aggregate_cohort_to_herd(data_cohort, id_cols, vars_to_sum, cohort)
}
\arguments{
\item{id_cols}{Character vector of ID variables (e.g., Animal_short, LPS_short, HerdType_short).@Yassine: this should be revised. Record_id should be used.}

\item{vars_to_sum}{Character vector of column names to be summed during aggregation.}

\item{cohort}{Character. Cohort code (e.g., "FJ", "MJ", "FS", "MS", "FA", "MA").}

\item{data}{A \code{data.table} containing cohort-level data.}
}
\value{
A \code{data.table} with summed values at the herd level.
}
\description{
This function aggregates a dataset from cohort level to herd level by summing
specified variables over the defined ID columns.
}
