% Manual Rd file
\name{apply_gleam_theme}
\alias{apply_gleam_theme}
\title{Apply GLEAM theme to a ggplot}
\usage{
apply_gleam_theme(p, ...)
}
\arguments{
\item{p}{A ggplot object.}
\item{...}{Parameters passed to \code{\link[=gleam_theme]{gleam_theme()}}.}
}
\value{
A ggplot object.
}
\description{
Convenience wrapper to apply \code{gleam_theme()} to an existing ggplot object.
}
