% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_utils.R
\name{as_geneset}
\alias{as_geneset}
\title{Coerce geneset input to named list}
\usage{
as_geneset(x)
}
\arguments{
\item{x}{Geneset input.}
}
\value{
Named list.
}
\description{
Coerce geneset input to named list
}
\keyword{internal}
