% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_model_weights.R
\name{calc_avg_weights}
\alias{calc_avg_weights}
\title{Calculate average and final live weights by cohort}
\usage{
calc_avg_weights(
  live_weight_cohort_initial,
  live_weight_cohort_potential_final,
  live_weight_cohort_at_slaughter,
  offtake_rate
)
}
\arguments{
\item{live_weight_cohort_initial}{Numeric. Live weight at the beginning of the cohort stage (kg).}

\item{live_weight_cohort_potential_final}{Numeric. Potential final live weight attainable at the end of the cohort stage in the absence of offtake (kg). (For juveniles: equals weaning weight; For subadults: equals adult live weight; For adults: equals adult live weight)}

\item{live_weight_cohort_at_slaughter}{Numeric. Live weight at slaughter for animals removed from the cohort (kg).}

\item{offtake_rate}{Numeric. Annual proportion of animals removed from the herd for each sex-age cohort (fraction).}
}
\value{
A named list with:
\describe{
\item{live_weight_cohort_average}{Numeric. Average live weight over the cohort stage. Computed by accounting for the share of offtaken animals within the cohort, using their slaughter weight, and the potential final weight of animals that remain in the cohort (kg).}
\item{live_weight_cohort_final}{Numeric. Live weight at the end of the cohort stage, accounting for both surviving and offtaken animals. Computed as a weighted average of the potential final weight of surviving animals and the slaughter weight of offtaken animals, based on the offtake rate (kg).}
}
}
\description{
Calculates the average and final live weight for a given
sex–age cohort based on initial weight,
potential final weight, slaughter weight, and the offtake rate.
}
\details{
The calculation of \code{live_weight_cohort_average} and \code{live_weight_cohort_final} is performed considering that
a fraction of animals is removed (offtake) during the cohort stage, while the
remaining animals reach the potential final live weight.

The final live weight is computed as:
\deqn{
live\_weight\_cohort_final =
(1 - offtake\_rate) \times live\_weight\_cohort\_potential\_final +
offtake\_rate \times live\_weight\_cohort\_at\_slaughter
}

The average live weight over the stage is approximated as:
\deqn{live\_weight\_cohort\_average = 
(live\_weight\_cohort\_initial + live\_weight\_cohort\_final)/2}

This function is part of the \code{\link[=run_weights_module]{run_weights_module()}}.
}
\seealso{
\code{\link{run_weights_module}}
}
