% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_simulation_core_model.R
\name{calc_cohort_weights}
\alias{calc_cohort_weights}
\title{Calculate Live Weights by Cohort and at different lifestage}
\usage{
calc_cohort_weights(
  animal,
  cohort,
  adult_female_weight = NA_real_,
  adult_male_weight = NA_real_,
  birth_weight = NA_real_,
  slaughter_weight_female = NA_real_,
  slaughter_weight_male = NA_real_,
  weaning_weight = NA_real_,
  age_first_calving = NA_real_,
  animal_age = NA_real_
)
}
\arguments{
\item{animal}{Character. Species code (e.g., "PGS", "CML", "CTL", "BFL", "SHP", "GTS").}

\item{cohort}{Character. Cohort code (e.g., "FJ", "MJ", "FS", "MS", "FA", "MA").}

\item{adult_female_weight}{Numeric. Adult female weight in kg.}

\item{adult_male_weight}{Numeric. Adult male weight in kg.}

\item{birth_weight}{Numeric. Birth weight in kg.}

\item{slaughter_weight_female}{Numeric. Slaughter weight of adult female in kg.}

\item{slaughter_weight_male}{Numeric. Slaughter weight of adult male in kg.}

\item{weaning_weight}{Numeric. Weaning weight in kg.}

\item{age_first_calving}{Numeric. Age at first calving (in days).}

\item{animal_age}{Numeric. Age of the animal at the current stage (in days).}
}
\value{
A named list with:
\describe{
\item{initial_weight}{Initial live weight.}
\item{potential_final_weight}{Potential final live weight.}
\item{slaughter_weight}{Slaughter live weight.}
}
}
\description{
Computes initial, potential final, and slaughter live weight (LW) for
a given cohort and animal type.
}
