% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_simulation_functions.R
\name{calc_cohort_weights}
\alias{calc_cohort_weights}
\title{Calculate Live Weights by Cohort and at different lifestage}
\usage{
calc_cohort_weights(
  Animal_short,
  cohort,
  AFKG = NA_real_,
  AMKG = NA_real_,
  CKG = NA_real_,
  MFSKG = NA_real_,
  MMSKG = NA_real_,
  WKG = NA_real_,
  AFC = NA_real_,
  WA = NA_real_
)
}
\arguments{
\item{Animal_short}{Character. Species code (e.g., "PGS", "CML", "CTL", "BFL", "SHP", "GTS").}

\item{cohort}{Character. Cohort code (e.g., "FJ", "MJ", "FS", "MS", "FA", "MA").}

\item{AFKG}{Numeric. Adult female weight in kg.}

\item{AMKG}{Numeric. Adult male weight in kg.}

\item{CKG}{Numeric. Birth weight in kg.}

\item{MFSKG}{Numeric. Slaughter weight of adult female in kg.}

\item{MMSKG}{Numeric. Slaughter weight of adult male in kg.}

\item{WKG}{Numeric. Weaning weight in kg.}

\item{AFC}{Numeric. Age at first calving (in days).}

\item{WA}{Numeric. Age of the animal at the current stage (in days).}
}
\value{
A named list with:
\describe{
\item{initial_weight}{Initial live weight.}
\item{potential_final_weight}{Potential final live weight.}
\item{slaughter_weight}{Slaughter live weight.}
}
}
\description{
Computes initial, potential final, and slaughter live weight (LW) for
a given cohort and animal type.
}
