% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_model_weights.R
\name{calc_daily_weight_gain}
\alias{calc_daily_weight_gain}
\title{Calculate daily weight gain by cohort}
\usage{
calc_daily_weight_gain(
  live_weight_cohort_potential_final,
  live_weight_cohort_initial,
  cohort_duration_days
)
}
\arguments{
\item{live_weight_cohort_potential_final}{Numeric. Potential final live weight attainable at the end of the cohort stage in the absence of offtake (kg). (For juveniles: equals weaning weight; For subadults: equals adult live weight; For adults: equals adult live weight)}

\item{live_weight_cohort_initial}{Numeric. Live weight at the beginning of the cohort stage (kg).}

\item{cohort_duration_days}{Numeric. Amount of time that each animal spends in a specific cohort (days). For \code{CHK}, \code{FJ} and \code{MJ} cohorts can default to 3 days.}
}
\value{
Numeric. Average live weight gain of the cohort over the cohort stage (kg/head/day).
}
\description{
Calculates average daily weight gain for a given
sex–age cohort based on the difference between potential final and initial live weights.
}
\details{
Daily live weight gain is calculated as the difference between the
potential final live weight and the initial live weight, divided by the
duration of the cohort stage:

\deqn{daily\_weight\_gain = 
(live\_weight\_cohort\_potential\_final - live\_weight\_cohort\_initial) / cohort\_duration\_days}

This function is part of the \code{\link[=run_weights_module]{run_weights_module()}}.
}
\seealso{
\code{\link{run_weights_module}}
}
