% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed_emissions_core_model.R
\name{calc_diet_co2_feed_luc_nopeat}
\alias{calc_diet_co2_feed_luc_nopeat}
\title{Calculate diet Carbon dioxide (CO₂) from land-use change (no peat) for feed production}
\usage{
calc_diet_co2_feed_luc_nopeat(feed_ration_fraction, co2_feed_luc_nopeat)
}
\arguments{
\item{feed_ration_fraction}{Numeric. Proportion of a specific feed item in the total ration, expressed as its fraction of diet dry matter (fraction). Within each herd_id and cohort, proportions should sum to 1.}

\item{co2_feed_luc_nopeat}{Numeric. Carbon dioxide (CO₂) emission factor from land-use change excluding peat in feed production, calculated per kg of dry matter intake (g CO₂/kg DM).}
}
\value{
Numeric. Average carbon dioxide (CO₂) emission factor from land-use change excluding peat in feed production of the diet (g CO₂/kg DM).
}
\description{
Computes the weighted contribution of Carbon dioxide (CO₂) emissions from land-use change
excluding peatland conversion associated with feed production for a feed ration component.
}
\details{
The contribution is computed as:

\deqn{diet\_co2\_feed\_luc\_nopeat = feed\_ration\_fraction \times co2\_feed\_luc\_nopeat}
}
