% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirect_emissions_feed_core_model.R
\name{calc_diet_n2o_feed_manure_applied}
\alias{calc_diet_n2o_feed_manure_applied}
\title{Calculate diet N2O from manure application (feed production)}
\usage{
calc_diet_n2o_feed_manure_applied(
  feed_ration_fraction,
  n2o_feed_manure_applied
)
}
\arguments{
\item{feed_ration_fraction}{Numeric. Fraction of the total ration represented
by this feed component (unitless).}

\item{n2o_feed_manure_applied}{Numeric. Emission factor for N2O from manure
application in feed production (g N2O/kg dry matter of feed).}
}
\value{
Numeric. N2O contribution from manure application for this feed item (g N2O/kg dry matter of feed).
}
\description{
Computes the weighted contribution of N2O emissions from manure applied to cropland
associated with feed production.
}
\details{
\deqn{diet\_n2o\_feed\_manure\_applied =
  feed\_ration\_fraction \times n2o\_feed\_manure\_applied}

This helper is vectorized over its inputs.
}
