% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_requirements_core_model.R
\name{calc_dry_matter_intake}
\alias{calc_dry_matter_intake}
\title{Calculate Daily Dry Matter Intake}
\usage{
calc_dry_matter_intake(
  species_short,
  energy_requirement_total,
  diet_gross_energy,
  diet_metabolizable_energy
)
}
\arguments{
\item{species_short}{Character. Code identifying the livestock species.
Supported values include:
\itemize{
\item \code{PGS}: pigs
\item \code{CML}: camels
\item \code{CTL}: cattle
\item \code{BFL}: buffalo
\item \code{SHP}: sheep
\item \code{GTS}: goats
}}

\item{energy_requirement_total}{Numeric. Total daily energy requirement (MJ/head/day).
For CTL, BFL, SHP and GTS this is expressed as gross energy intake requirement (GE).
For CML and PGS the function returns the summed daily metabolizable energy requirement.}

\item{diet_gross_energy}{Numeric. Average gross energy content of the diet (MJ/kg DM).}

\item{diet_metabolizable_energy}{Numeric. Average metabolizable energy content of the diet (MJ/kg DM).}
}
\value{
Numeric. Average daily dry matter intake of feed (kg DM/head/day).
}
\description{
Computes daily feed intake as dry matter intake (DMI) per animal (kg DM/head/day) from the
animal's daily energy requirement and the diet energy density.
}
\details{
This function follows the IPCC Tier 2, logic.
Dry matter intake is calculated by dividing the appropriate daily energy requirement by the
corresponding diet energy density (MJ per kg DM).

This function follows the IPCC Tier 2 framework.
DMI is computed
by dividing the appropriate daily energy requirement by the corresponding
diet energy density (MJ/kg DM).

\itemize{
\item \strong{Energy expressed as gross energy intake requirement -
\code{CTL}, \code{BFL}, \code{SHP}, \code{GTS}}:

\deqn{
    dry\_matter\_intake = \frac{energy\_requirement\_total}{diet\_gross\_energy}
  }

\item \strong{Energy expressed as metabolizable energy requirement -
\code{CML}, \code{PGS}, \code{CHK}}:

\deqn{
    dry\_matter\_intake = \frac{energy\_requirement\_total}{diet\_metabolizable\_energy}
  }
}
}
\references{
IPCC. (2006). \emph{2006 IPCC Guidelines for National Greenhouse Gas Inventories},
Volume 4 (AFOLU), Chapter 10: \emph{Emissions from Livestock and Manure Management}.

IPCC. (2019). \emph{2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories},
Volume 4 (AFOLU), Chapter 10: \emph{Emissions from Livestock and Manure Management}.
}
\seealso{
\code{\link{calc_total_energy_requirement}}
\code{\link{calc_diet_gross_energy}}
\code{\link{calc_diet_metabolizable_energy}}
}
