% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation_core_model.R
\name{calc_energy_allocation_milk}
\alias{calc_energy_allocation_milk}
\title{Milk Energy Requirement for Allocation}
\usage{
calc_energy_allocation_milk(
  milk_fpcm_output,
  standard_protein,
  standard_fat,
  standard_lactose
)
}
\arguments{
\item{milk_fpcm_output}{Numeric scalar. Total Fat-protein-corrected milk (FPCM) produced over the assessment period  (kg/assessment period). Default fat and protein content=0.04 and 0.033.}

\item{standard_protein}{Numeric scalar. Reference protein content (g per 100 g milk).}

\item{standard_fat}{Numeric scalar. Reference fat content (g per 100 g milk).}

\item{standard_lactose}{Numeric scalar. Reference lactose content (g per 100 g milk).}
}
\value{
Numeric scalar. Energy requirements in megajoules.
}
\description{
Converts fat- and protein-corrected milk output into the megajoule demand used in the allocation workflow.
The formula calculates energy density based on standard milk composition and multiplies by production.
}
