% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production_cohort_core_model.R
\name{calc_meat_production}
\alias{calc_meat_production}
\title{Compute Meat Production Outputs}
\usage{
calc_meat_production(
  offtake_heads_assessment,
  slaughter_weight_cohort,
  carcass_dressing_fraction,
  bone_free_meat_fraction,
  meat_protein_fraction
)
}
\arguments{
\item{offtake_heads_assessment}{Numeric. Total number of animals removed via offtake over the assessment period,
aggregated to 6 sex–age cohorts (heads/assessment period) (cohorts = FJ, FS, FA, MJ, MS, MA).}

\item{slaughter_weight_cohort}{Numeric. Live weight at slaughter for animals removed from the cohort (kg).}

\item{carcass_dressing_fraction}{Numeric. Ratio of a slaughtered animal's carcass weight to its live weight
(fraction).}

\item{bone_free_meat_fraction}{Numeric. Ratio of bone-free-meat to carcass weight (fraction).}

\item{meat_protein_fraction}{Numeric. Protein content of bone-free-meat (kg protein/kg bone-free-meat).}
}
\value{
A named list with:
\describe{
\item{meat_production_live_weight_cohort}{Numeric . Total meat produced as live weight over the assessment period by
cohort (kg/cohort/assessment period).}
\item{meat_production_carcass_weight_cohort}{Numeric. Total meat as carcass weight (excluding organs, and other
by-products after dressing) produced over the assessment period by cohort (kg/cohort/assessment period).}
\item{meat_production_bone_free_meat_cohort}{Numeric. Total bone-free-meat (excluding bones, organs, and other
by-products after dressing and bone removal)
produced over the assessment period by cohort (kg/cohort/assessment period).}
\item{meat_production_protein_cohort}{Numeric. Total meat protein (excluding bones, organs, and other by-products
after dressing and bone removal) produced
over the assessment period by cohort (kg protein/cohort/assessment period).}
}
}
\description{
Computes cohort-level meat production outputs over the assessment period based on
the number of animals removed from the herd (offtake). The function returns
multiple production metrics expressed in live weight, carcass weight,
bone-free meat, and meat protein (kg/cohort/assessment period).
}
\details{
Meat production outputs are computed as follows:
\itemize{
\item \strong{\code{meat_production_live_weight_cohort}} is computed as:

\eqn{meat\_production\_live\_weight\_cohort =
  offtake\_heads\_assessment \times slaughter\_weight\_cohort}

\item \strong{\code{meat_production_carcass_weight_cohort}} is computed as:

\eqn{meat\_production\_carcass\_weight\_cohort =
  meat\_production\_live\_weight\_cohort \times carcass\_dressing\_fraction}

\item \strong{\code{meat_production_bone_free_meat_cohort}} is computed as:

\eqn{meat\_production\_bone\_free\_meat\_cohort =
  meat\_production\_carcass\_weight\_cohort \times bone\_free\_meat\_fraction}

\item \strong{\code{meat_production_protein_cohort}} is computed as:

\eqn{meat\_production\_protein\_cohort =
  meat\_production\_bone\_free\_meat\_cohort \times meat\_protein\_fraction}
}
}
\seealso{
\code{\link{run_demographic_herd_module}}
\code{\link{run_weights_module}}
}
