% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directemissions_manure_core_model.R
\name{calc_n2o_manure_total}
\alias{calc_n2o_manure_total}
\title{Calculate total N₂O emissions from manure}
\usage{
calc_n2o_manure_total(
  n2o_vol_manure_pasture,
  n2o_leach_manure_pasture,
  n2o_vol_manure_burned,
  n2o_leach_manure_burned,
  n2o_vol_manure_other,
  n2o_leach_manure_other,
  n2o_manure_pasture_direct,
  n2o_manure_burned_direct,
  n2o_manure_other_direct
)
}
\arguments{
\item{n2o_vol_manure_pasture}{Numeric. Indirect nitrous oxide (N₂O) emissions
resulting from atmospheric deposition of volatilised nitrogen (NH₃ and NOₓ)
from manure deposited on pasture (kg N₂O/head/day).}

\item{n2o_leach_manure_pasture}{Numeric. Indirect nitrous oxide (N₂O) emissions
resulting from leaching and runoff of manure nitrogen from manure deposited on
pasture (kg N₂O/head/day).}

\item{n2o_vol_manure_burned}{Numeric. Indirect nitrous oxide (N₂O) emissions
resulting from atmospheric deposition of volatilised nitrogen (NH₃ and NOₓ)
from manure burned for fuel (kg N₂O/head/day).}

\item{n2o_leach_manure_burned}{Numeric. Indirect nitrous oxide (N₂O) emissions
resulting from leaching and runoff of manure nitrogen from manure burned for
fuel (kg N₂O/head/day).}

\item{n2o_vol_manure_other}{Numeric. Indirect nitrous oxide (N₂O) emissions
resulting from atmospheric deposition of volatilised nitrogen (NH₃ and NOₓ)
from manure management systems, excluding manure deposited on pasture and
manure burned for fuel (kg N₂O/head/day).}

\item{n2o_leach_manure_other}{Numeric. Indirect nitrous oxide (N₂O) emissions
resulting from leaching and runoff of manure nitrogen from manure management
systems, excluding losses from manure deposited on pasture and manure burned
for fuel (kg N₂O/head/day).}

\item{n2o_manure_pasture_direct}{Numeric. Direct nitrous oxide (N₂O) emissions
from manure deposited on pasture (kg N₂O/head/day).}

\item{n2o_manure_burned_direct}{Numeric. Direct nitrous oxide (N₂O) emissions
from manure burned for fuel (kg N₂O/head/day).}

\item{n2o_manure_other_direct}{Numeric. Direct nitrous oxide (N₂O) emissions
from manure management systems, excluding emissions from manure deposited on
pasture and burned for fuel (kg N₂O/head/day).}
}
\value{
A named list with:
\describe{
\item{n2o_manure_pasture_indirect}{
Numeric. Total indirect nitrous oxide (N₂O) emissions from manure deposited
on pasture. Includes emissions from atmospheric deposition of volatilised
nitrogen (NH₃ and NOₓ) and from leaching and runoff of manure nitrogen
(kg N₂O/head/day).
}
\item{n2o_manure_burned_indirect}{
Numeric. Total indirect nitrous oxide (N₂O) emissions originating from
manure burned for fuel. Includes emissions from atmospheric deposition of
volatilised nitrogen (NH₃ and NOₓ) and from leaching and runoff of manure
nitrogen (kg N₂O/head/day).
}
\item{n2o_manure_other_indirect}{
Numeric. Total indirect nitrous oxide (N₂O) emissions originating from
manure management systems, excluding manure deposited on pasture and
manure burned for fuel. Includes emissions from atmospheric deposition of
volatilised nitrogen (NH₃ and NOₓ) and from leaching and runoff of manure
nitrogen (kg N₂O/head/day).
}
\item{n2o_manure_pasture_total}{
Numeric. Total nitrous oxide (N₂O) emissions from manure deposited on
pasture. Includes direct emissions and indirect emissions from
volatilisation, leaching, and runoff (kg N₂O/head/day).
}
\item{n2o_manure_burned_total}{
Numeric. Total nitrous oxide (N₂O) emissions from manure burned for fuel.
Includes direct emissions and indirect emissions from volatilisation,
leaching, and runoff (kg N₂O/head/day).
}
\item{n2o_manure_other_total}{
Numeric. Total nitrous oxide (N₂O) emissions from manure management
systems, excluding manure deposited on pasture and manure burned for fuel.
Includes direct emissions and indirect emissions from volatilisation,
leaching, and runoff (kg N₂O/head/day).
}
}
}
\description{
Aggregates direct and indirect nitrous oxide (N₂O) emissions from manure, by
manure management system group (deposited on pasture, burned for fuel, and all other systems).
Indirect emissions include contributions from volatilization and from leaching
and runoff.
}
\details{
The following aggregations are applied:
\deqn{
  n2o\_manure\_pasture\_indirect =
  n2o\_vol\_manure\_pasture + n2o\_leach\_manure\_pasture
}
\deqn{
  n2o\_manure\_burned\_indirect =
  n2o\_vol\_manure\_burned + n2o\_leach\_manure\_burned
}
\deqn{
  n2o\_manure\_other\_indirect =
  n2o\_vol\_manure\_other + n2o\_leach\_manure\_other
}
\deqn{
  n2o\_manure\_pasture\_total =
  n2o\_manure\_pasture\_indirect + n2o\_manure\_pasture\_direct
}
\deqn{
  n2o\_manure\_burned\_total =
  n2o\_manure\_burned\_indirect + n2o\_manure\_burned\_direct
}
\deqn{
  n2o\_manure\_other\_total =
  n2o\_manure\_other\_indirect + n2o\_manure\_other\_direct
}
}
\examples{
calc_n2o_manure_total(
  n2o_vol_manure_pasture = 0.0129,
  n2o_leach_manure_pasture = 0.0012,
  n2o_vol_manure_burned = 0,
  n2o_leach_manure_burned = 0,
  n2o_vol_manure_other = 0.052,
  n2o_leach_manure_other = 0.00027,
  n2o_manure_pasture_direct = 0.009,
  n2o_manure_burned_direct = 0,
  n2o_manure_other_direct = 0.01033
)

}
