% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_requirements_core_model.R
\name{calc_net_energy_activity}
\alias{calc_net_energy_activity}
\title{Calculate Energy for Activity}
\usage{
calc_net_energy_activity(
  species_short,
  cohort_short,
  energy_requirement_maintenance,
  live_weight_cohort_average,
  low_activity_fraction,
  high_activity_fraction
)
}
\arguments{
\item{species_short}{Character. Code identifying the livestock species.
Supported values include:
\itemize{
\item \code{PGS}: pigs
\item \code{CML}: camels
\item \code{CTL}: cattle
\item \code{BFL}: buffalo
\item \code{SHP}: sheep
\item \code{GTS}: goats
}}

\item{cohort_short}{Character. Sex- and age-specific cohort code describing the
production stage of the animals. Supported values include:
\itemize{
\item \code{FA}: adult females (from age at first parturition)
\item \code{FS}: sub-adult females (from weaning to age at first parturition)
\item \code{FJ}: juvenile females (from birth to weaning)
\item \code{MA}: adult males (from age at first breeding)
\item \code{MS}: sub-adult males (from weaning to age at first breeding)
\item \code{MJ}: juvenile males (from birth to weaning)
}}

\item{energy_requirement_maintenance}{Numeric. Energy required for maintenance, defined as the amount of energy needed to
keep the animal at equilibrium such that body energy is neither gained nor lost.
Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for
CML and PGS (MJ/head/day).}

\item{live_weight_cohort_average}{Numeric. Average live weight over the cohort stage. Computed by accounting
for the share of offtaken animals within the cohort, using their slaughter weight, and the potential
final weight of animals that remain in the cohort (kg).}

\item{low_activity_fraction}{Numeric. Proportion of the assessment period during which the animal performs
low-intensity movement typical of stall-feeding or near-field grazing, characterized by minimal walking distances and flat terrain  (fraction).}

\item{high_activity_fraction}{Numeric. Proportion of the assessment period during which the animal engages in sustained locomotion associated
with herding or long-distance grazing, typically involving extended walking distances and/or uneven or hilly terrain (fraction).}
}
\value{
Numeric. Energy required for activity, defined as the amount of energy needed to support animal movement and physical activity (MJ/head/day).
Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS.
}
\description{
Computes the energy requirement for activity by cohort (MJ/head/day),
defined as the amount of energy needed to support animal movement and physical activity.
}
\details{
This approach follows the IPCC Tier 2 energy partitioning method and applies:

\eqn{energy\_requirement\_activity = cact \times energy\_requirement\_maintenance}

For \code{SHP} and \code{GTS}, activity energy is calculated as:

\eqn{energy\_requirement\_activity = cact \times live\_weight\_cohort\_average}

where

\eqn{cact} is an activity coefficient (dimensionless for \code{CTL}, \code{BFL}, \code{PGS};
MJ/day/kg for \code{SHP},  \code{GTS}) that reflects the animal’s feeding and management conditions.
\code{energy_requirement_maintenance} can be calculated using
\code{\link{calc_net_energy_maintenance}()}.

For \code{CTL, BFL, SHP} and \code{GTS}, \eqn{cact} is computed as a \emph{weighted average} of activity levels
over the assessment period to account for variation in management and grazing intensity.

\strong{Specific coefficients by species and cohort:}

\strong{CTL, BFL} (NRC, 1996; AFRC, 1993):
\itemize{
\item \code{low_activity_fraction}: \eqn{cact = 0.17}
\item \code{high_activity_fraction}: \eqn{cact = 0.36}
}
\strong{CML} (Wardeh, 2004):
\itemize{
\item \code{all activity levels}: \eqn{cact = 0.1}
}
\strong{GTS} (AFRC, 1993):
\itemize{
\item \code{low_activity_fraction}: \eqn{cact = 0.019}
\item \code{high_activity_fraction}: \eqn{cact = 0.024}
}
\strong{SHP} (AFRC, 1993):
\itemize{
\item \code{low_activity_fraction}: \eqn{cact = 0.0107}
\item \code{high_activity_fraction}: \eqn{cact = 0.024}
}
\strong{PGS} (NRC, 1998):
\itemize{
\item \code{all activity levels}: \eqn{cact = 0.125}
}
}
\references{
NRC (1998). \emph{Nutrient Requirements of Swine},
10th Revised Edition. National Academies Press, Washington, DC.

NRC (1996). \emph{Nutrient Requirements of Beef Cattle},
7th Revised Edition. National Academies Press, Washington, DC.

AFRC (1993). \emph{Energy and Protein Requirements of Ruminants. An Advisory Manual Prepared by the AFRC Technical Committee on Responses to Nutrients.}
CAB International, Wallingford, UK.

IPCC. (2019). \emph{2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories},
Chapter 10: Emissions from Livestock and Manure Management, Equation 10.4; Table 10.5.

IPCC. (2006). \emph{2006 IPCC Guidelines for National Greenhouse Gas Inventories}, Chapter 10: Emissions
from Livestock and Manure Management, Equation 10.4; Table 10.5.

Wardeh, M. F. (2004). \emph{The nutrient requirements of the dromedary camel}.
Journal of Camel Science, 1(1):37-45. The Camel Applied Research and Development Network (CARDN),
Arab Center for the Studies of Arid Zones and Dry Lands (ACSAD).
}
\seealso{
\code{\link{calc_net_energy_maintenance}}
\code{\link{calc_avg_weights}}
}
