% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_requirements_core_model.R
\name{calc_net_energy_activity}
\alias{calc_net_energy_activity}
\title{Calculate Net Energy for Activity}
\usage{
calc_net_energy_activity(
  animal,
  cohort,
  nemain,
  average_weight,
  offtake_rate,
  activity_fraction,
  high_activity_fraction
)
}
\arguments{
\item{animal}{Character. Species code.}

\item{cohort}{Character. Cohort code.}

\item{nemain}{Numeric. Net energy for maintenance.}

\item{average_weight}{Numeric. Average live weight (kg).}

\item{offtake_rate}{Numeric. Offtake rate by cohort.}

\item{activity_fraction}{Numeric. Fraction of time under low-activity conditions.}

\item{high_activity_fraction}{Numeric. Fraction of time under high-activity conditions.}
}
\value{
Numeric. Net energy for activity (MJ/head/day).
}
\description{
Computes net energy required for activity (MJ/head/day).
}
