% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_requirements_core_model.R
\name{calc_net_energy_fibre}
\alias{calc_net_energy_fibre}
\title{Calculate Net Energy for Fibre Production}
\usage{
calc_net_energy_fibre(animal, cohort, fibre_prod)
}
\arguments{
\item{animal}{Character. Species code.}

\item{cohort}{Character. Cohort code.}

\item{fibre_prod}{Numeric. Fibre production (kg/year).}
}
\value{
Numeric. Net energy for fibre production (MJ/head/day).
}
\description{
Computes net energy required for fibre production (MJ/head/day).
}
