% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_requirements_core_model.R
\name{calc_net_energy_lactation}
\alias{calc_net_energy_lactation}
\title{Calculate Net Energy for Lactation}
\usage{
calc_net_energy_lactation(
  animal,
  cohort,
  milking_fraction,
  milk_yield,
  milk_fat,
  idle,
  gest,
  litsize,
  dr1,
  ckg,
  wkg,
  lact,
  parturition_rate,
  lambing_interval
)
}
\arguments{
\item{animal}{Character. Species code.}

\item{cohort}{Character. Cohort code.}

\item{milking_fraction}{Numeric. Proportion of lactating adult females.}

\item{milk_yield}{Numeric. Milk yield (kg/day).}

\item{milk_fat}{Numeric. Milk fat content (fraction).}

\item{idle}{Numeric. Fraction of time idle (for PGS).}

\item{gest}{Numeric. Fraction of time gestating (for PGS).}

\item{litsize}{Numeric. Litter size (for PGS, SHP, GTS).}

\item{dr1}{Numeric. Death rate in first year (for PGS).}

\item{ckg}{Numeric. Birth weight (for PGS).}

\item{wkg}{Numeric. Weaning weight (for PGS).}

\item{lact}{Numeric. Fraction of time lactating (for PGS).}

\item{parturition_rate}{Numeric. Parturition rate.}

\item{lambing_interval}{Numeric. Lambing interval (for SHP, GTS).}
}
\value{
Numeric. Net energy for lactation (MJ/head/day).
}
\description{
Computes net energy required for lactation (MJ/head/day).
}
