% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_requirements_core_model.R
\name{calc_net_energy_pregnancy}
\alias{calc_net_energy_pregnancy}
\title{Calculate Net Energy for Pregnancy}
\usage{
calc_net_energy_pregnancy(
  animal,
  cohort,
  nemain,
  parturition_rate,
  litsize,
  gest,
  duration,
  offtake_rate
)
}
\arguments{
\item{animal}{Character. Species code.}

\item{cohort}{Character. Cohort code.}

\item{nemain}{Numeric. Net energy for maintenance.}

\item{parturition_rate}{Numeric. Parturition rate.}

\item{litsize}{Numeric. Litter size (for SHP, GTS, PGS).}

\item{gest}{Numeric. Fraction of time gestating (for all species).}

\item{duration}{Numeric. Duration in days.}

\item{offtake_rate}{Numeric. Offtake rate by cohort.}
}
\value{
Numeric. Net energy for pregnancy (MJ/head/day).
}
\description{
Computes net energy required for pregnancy (MJ/head/day).
}
