% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrogen_balance_core_model.R
\name{calc_nitrogen_intake}
\alias{calc_nitrogen_intake}
\title{Compute Daily Nitrogen Intake}
\usage{
calc_nitrogen_intake(dry_matter_intake, ration_nitrogen)
}
\arguments{
\item{dry_matter_intake}{Numeric. Average daily dry matter intake of feed (kg DM/head/day).}

\item{ration_nitrogen}{Numeric. Average nitrogen content of diet (kg N/kg DM).}
}
\value{
Numeric. Daily nitrogen intake (kg N/head/day).
}
\description{
Calculates the daily nitrogen intake per head (kg N/head/day) as the product of feed dry matter
intake (DMI) and diet nitrogen content.
}
\details{
This approach follows the IPCC Tier 2 approach and estimates \code{dry_matter_intake}  as follows:

\eqn{nitrogen\_intake = dry\_matter\_intake \times diet\_nitrogen}
}
\references{
IPCC. (2019). \emph{2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories},
Chapter 10: Emissions from
Livestock and Manure Management. Equation 10.32.

IPCC. (2006). \emph{2006 IPCC Guidelines for National Greenhouse Gas Inventories},
Chapter 10: Emissions from
Livestock and Manure Management. Equation 10.32.
}
\seealso{
\code{\link{calc_ration_intake}},
\code{\link{run_metabolic_energy_req_module}},
\code{\link{calc_ration_nitrogen_content}},
\code{\link{run_ration_quality_module}}
}
