% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directemissions_manure_core_model.R
\name{calc_nitrogen_leaching}
\alias{calc_nitrogen_leaching}
\title{Calculate Nitrogen Leaching}
\usage{
calc_nitrogen_leaching(
  n_excretion,
  fracleach_pasture,
  fracleach_burned,
  fracleach_other
)
}
\arguments{
\item{n_excretion}{Numeric. Daily nitrogen excretion (kg N/head/day)}

\item{fracleach_pasture}{Numeric. Effective fraction of excreted nitrogen lost through leaching and runoff from manure deposited on pasture, already weighted by the share of manure deposited on pasture (mmspasture) (fraction).}

\item{fracleach_burned}{Numeric. Effective fraction of excreted nitrogen lost through leaching and runoff from manure burned for fuel, already weighted by the share of manure burned (mmsburned) (fraction).}

\item{fracleach_other}{Numeric. Effective fraction of excreted nitrogen lost through leaching and runoff from manure managed in non-pasture, non-burned manure management systems, already weighted by the shares of the corresponding manure management systems (e.g. solid storage, drylot, liquid/slurry) (fraction).}
}
\value{
A named list with:
\describe{
\item{n_leach_manure_pasture}{Numeric. Amount of manure nitrogen lost through leaching and runoff from manure deposited on pasture (kg N/head/day).}
\item{n_leach_manure_burned}{Numeric. Amount of manure nitrogen lost through leaching and runoff from manure burned for fuel (kg N/head/day).}
\item{n_leach_manure_other}{Numeric. Amount of manure nitrogen lost through leaching and runoff from manure managed in all other systems, excluding losses from manure deposited on pasture and manure burned for fuel (kg N/head/day).}
\item{n_leach_manure_all_noburn}{Numeric. Amount of manure nitrogen lost through leaching and runoff from manure managed in all other systems and deposited on pasture, excluding losses from manure burned for fuel (kg N/head/day).}
}
}
\description{
Calculates the amount of manure nitrogen lost via \strong{leaching and runoff}
from manure management systems (kg N/head/day). This follows the IPCC
structure of Eq. 10.27, but expressed on a \strong{daily} basis because
\code{n_excretion} is provided in kg N/head/day.
}
\details{
The function assumes leaching/runoff fractions - (\code{fracleach_pasture}, \code{fracleach_burned}, \code{fracleach_other}) -
are provided as \strong{effective
weighted fractions}, i.e. they have already been multiplied by the
corresponding manure management system shares (mms fractions).
}
\references{
IPCC. (2019). \emph{2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories}, Chapter 10: Emissions from
Livestock and Manure Management. Equation 10.27.

IPCC. (2006). \emph{2006 IPCC Guidelines for National Greenhouse Gas Inventories}, Chapter 10: Emissions from
Livestock and Manure Management. Equation 10.27.
}
