% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_model_nondemographic_herd.R
\name{calc_nondemo_start_sizes}
\alias{calc_nondemo_start_sizes}
\title{Calculate nondemographic entrants across production cycles}
\usage{
calc_nondemo_start_sizes(
  cohort_stock_nondemo_annual_entrants,
  total_nondemo_cycle_starts_to_distribute
)
}
\arguments{
\item{cohort_stock_nondemo_annual_entrants}{Numeric. Total number of animals
entering the nondemographic production pathway over the simulated period
for the assessed cohort block (\code{FN} or \code{MN})
(\# heads / simulated period).}

\item{total_nondemo_cycle_starts_to_distribute}{Integer. Total number of
cycle starts within the fixed 365-day time horizon for the assessed
nondemographic cohort (\code{FN} or \code{MN}) (cycle starts / simulated
period).}
}
\value{
A named list with the following element:
\describe{
\item{cohort_stock_nondemo_start_cycle}{Numeric. Number of animals starting
each nondemographic production cycle for the assessed cohort block
(\code{FN} or \code{MN}), calculated by distributing total annual entrants
over the total number of cycle starts (\# heads / cycle).}
}
}
\description{
Calculates the total annual number of animals entering a nondemographic cohort
into the number of animals entering each production cycle, based on the number
of cycle starts occurring within the fixed 365-day horizon.
}
\details{
The annual entrant stock is evenly distributed across all defined production
cycle starts. The number of animals entering each cycle is calculated as:

\deqn{
cohort\_stock\_nondemo\_start\_cycle =
\frac{cohort\_stock\_nondemo\_annual\_entrants}
{total\_non\_demo\_cycle\_starts\_to\_distribute}
}

where:
\itemize{
\item \code{cohort\_stock\_nondemo\_start\_cycle} is the number of animals entering each cycle
\item \code{cohort\_stock\_nondemo\_annual\_entrants} is the total annual entrants
\item \code{total\_non\_demo\_cycle\_starts\_to\_distribute} is the number of cycle starts
}

If \code{total\_nondemo\_cycle\_starts\_to\_distribute} is \code{NULL} or less than or equal
to zero, the function assumes that no cycle structure is defined. In this case,
all annual entrants are assigned to a single implicit cycle:

\deqn{
cohort\_stock\_nondemo\_start\_cycle =
cohort\_stock\_nondemo\_annual\_entrants
}

No rounding or integer enforcement is applied:

This means fractional values may occur when
\code{cohort\_stock\_nondemo\_annual\_entrants} is not perfectly divisible
by \code{total\_non\_demo\_cycle\_starts\_to\_distribute}. Handling of integer constraints
(if required) is delegated to downstream processes.

This function is part of the \code{\link[=run_nondemographic_herd_module]{run_nondemographic_herd_module()}}.
}
\seealso{
\code{\link[=run_nondemographic_herd_module]{run_nondemographic_herd_module()}}
}
