% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_model_metabolic_energy_req.R
\name{calc_rem_maintenance}
\alias{calc_rem_maintenance}
\title{Calculate ratio of net energy available for maintenance in the diet (REM - Net Energy for Maintenance / Digestible Energy)}
\usage{
calc_rem_maintenance(species_short, ration_digestibility_fraction = NA_real_)
}
\arguments{
\item{species_short}{Character. Code identifying the livestock species.
Supported values include:
\itemize{
\item \code{PGS}: pigs
\item \code{CML}: camels
\item \code{CTL}: cattle
\item \code{BFL}: buffalo
\item \code{SHP}: sheep
\item \code{GTS}: goats
}}

\item{ration_digestibility_fraction}{Numeric. Average digestibility of
the feed ration, expressed as ratio of digestible to gross energy content (fraction).}
}
\value{
Numeric. Ratio of net energy available for maintenance in the diet to digestible energy consumed (fraction).
}
\description{
Calculates the ratio of net energy available in the diet for maintenance to digestible energy.
}
\details{
This component follows the IPCC Tier 2 partitioning approach and it returns the value for
ruminants (\code{CTL}, \code{BFL}, \code{SHP}, \code{GTS}) calculated as follows:

\deqn{
net\_energy\_maintenance\_digestible\_energy\_ratio =
1.123 - 0.004092 \times (diet\_digestibility\_fraction \times 100) +
0.00001126 \times (diet\_digestibility\_fraction \times 100)^2 -
\frac{25.4}{diet\_digestibility\_fraction \times 100}
}

For the Other species REM is not applicable and the function
returns \code{NA_real_}.

This function is part of the \code{\link[=run_metabolic_energy_req_module]{run_metabolic_energy_req_module()}}.
}
\references{
Gibbs M.J., Johnson D.E. (1993) \emph{Livestock Emissions}.
In: International Methane Emissions. Washington, D.C., U.S.A: US Environmental Protection Agency,
Climate Change Division.

IPCC. (2019). \emph{2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories},
Chapter 10: Emissions from Livestock and Manure Management, Equation 10.14.

IPCC. (2006). \emph{2006 IPCC Guidelines for National Greenhouse Gas Inventories}, Chapter 10: Emissions
from Livestock and Manure Management, Equation 10.14.
}
\seealso{
\code{\link{run_metabolic_energy_req_module}},
\code{\link{calc_ration_digestibility}}
\code{\link{calc_total_metabolic_energy_req}}
\code{\link{calc_ration_intake}}
}
