% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directemissions_manure_core_model.R
\name{calc_volatile_solids}
\alias{calc_volatile_solids}
\title{Calculate Volatile Solids for Manure Emissions}
\usage{
calc_volatile_solids(
  dry_matter_intake,
  diet_digestibility_fraction,
  urinary_energy_fraction = 0.04,
  diet_ash = 0.08
)
}
\arguments{
\item{dry_matter_intake}{Numeric. Daily dry matter intake of feed (kg DM/head/day).}

\item{diet_digestibility_fraction}{Numeric. Average digestibility of the the feed ration, expressed as ratio of digestible to gross energy content (fraction)}

\item{urinary_energy_fraction}{Numeric. Average ash content of feed, calculated as a fraction of the dry matter intake (kg ash/kg DM)}

\item{diet_ash}{Numeric. Fraction of animal's gross energy that is excreted in urine (fraction).}
}
\value{
Numeric. Total volatile solids (volatile_solids) excreted per animal per day, representing the organic material in livestock manure and consisting of both biodegradable and non-biodegradable fractions (kg VS/head/day).
}
\description{
Computes daily volatile solids (VS) excretion in manure (kg VS/head/day).
VS represents the total organic material excreted (biodegradable + non-biodegradable)
and is required to proceed with the estimate of methane emissions from manure management.
}
\details{
The IPCC recommends estimating VS from feed intake and digestibility when
country-specific average daily VS excretion rates are not available. The core relationship is
given in \strong{IPCC Equation 10.24 (Volatile solids excretion rates)}, which uses gross energy (GE)
intake, digestibility (DE), urinary energy as a fraction of GE (UE·GE), ash fraction (ASH), and
a conversion factor of 18.45 MJ/kg DM.

\strong{Implementation note (simplified coefficients).}
This package uses simplified algebraic forms by species/method that are consistent with the
structure of Eq. 10.24.
Specifically, in IPCC guidelines the the average gross energy content of the ration is used instead
of a fixed value of 18.45 MJ×kg DM-1. Thus, ge / diet_ge equals the daily intake, dmi.
}
\examples{
calc_volatile_solids <- calc_volatile_solids(
  dry_matter_intake = 5,
  diet_digestibility_fraction = 0.6,
  urinary_energy_fraction = 0.04,
  diet_ash = 0.08
)

}
\references{
IPCC. (2019). \emph{2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories}, Chapter 10: Emissions from
Livestock and Manure Management. Equation 10.24.

IPCC. (2006). \emph{2006 IPCC Guidelines for National Greenhouse Gas Inventories}, Chapter 10: Emissions from
Livestock and Manure Management. Equation 10.24.
}
