% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directemissions_enteric_core_model.R
\name{compute_daily_enteric_emissions}
\alias{compute_daily_enteric_emissions}
\title{Compute Daily Enteric Methane Emissions}
\usage{
compute_daily_enteric_emissions(animal, cohort, ym, diet_ge, dmi)
}
\arguments{
\item{animal}{Character. Species code  (e.g., \code{PGS}, \code{CML}, \code{CTL}, \code{BFL}, \code{SHP}, \code{GTS}).}

\item{cohort}{Character. Cohort code (e.g., \code{FA}, \code{FS}, \code{FJ}, \code{MA}, \code{MS},\code{MJ}).}

\item{ym}{Numeric. Methane conversion factor (YM), representing the share of  gross energy of the feed ration that is converted to CH₄.}

\item{diet_ge}{Numeric. Average gross energy content of the diet (MJ/kg DM).}

\item{dmi}{Numeric. Daily dry matter intake of feed (kg DM/head/day).}
}
\value{
Numeric scalar. Daily enteric methane emissions (kg CH₄/head/day).
}
\description{
Calculates daily enteric methane emissions (kg CH₄/head/day) based on gross energy intake,
methane conversion factor (YM), and dry matter intake (DMI). The formula assumes:
}
\details{
\deqn{CH₄ = diet\_ge × dmi × (ym / 100) / 55.65}

where 55.65 MJ/kg is the energy content of methane. Returns \code{NA} for chickens.
}
\references{
IPCC. (2019). \emph{2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories}, Chapter 10: Emissions from
Livestock and Manure Management, Equation 10.21.

IPCC. (2006). \emph{2006 IPCC Guidelines for National Greenhouse Gas Inventories}, Chapter 10: Emissions from
Livestock and Manure Management, Equation 10.21.
}
