% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directemissions_enteric_core_model.R
\name{compute_daily_enteric_emissions}
\alias{compute_daily_enteric_emissions}
\title{Compute Daily Enteric Methane Emissions}
\usage{
compute_daily_enteric_emissions(
  species_short,
  ch4_conversion_factor_ym,
  ch4_mitigation_factor,
  diet_gross_energy,
  dry_matter_intake
)
}
\arguments{
\item{species_short}{Character. Code identifying the livestock species.
Supported values include:
\itemize{
\item \code{PGS}: pigs
\item \code{CML}: camels
\item \code{CTL}: cattle
\item \code{BFL}: buffalo
\item \code{SHP}: sheep
\item \code{GTS}: goats
}}

\item{ch4_conversion_factor_ym}{Numeric. Methane conversion factor (ym),
representing the percentage of gross energy of the feed ration that is
converted to CH₄ (percentage).}

\item{ch4_mitigation_factor}{Numeric. Dimensionless fraction of baseline
enteric methane emissions remaining after mitigation (1 = no mitigation,
0.9 = 10\% reduction). Default = 1.}

\item{diet_gross_energy}{Numeric. Average gross energy content of the diet
(MJ/kg DM).}

\item{dry_matter_intake}{Numeric. Average daily dry matter intake of feed
(kg DM/head/day).}
}
\value{
Numeric. Average daily enteric methane emissions (kg CH₄/head/day).
}
\description{
Calculates daily enteric methane emissions (kg CH₄/head/day) based on gross
energy intake, methane conversion factor (ym), and dry matter intake.
}
\details{
The formula used to estimate daily enteric methane emissions is:

\deqn{CH_4 = \frac{diet\_gross\_energy \times dry\_matter\_intake \times \frac{ch4\_conversion\_factor\_ym}{100}}{55.65}}

where 55.65 MJ/kg is the energy content of methane.

The function returns \code{NA} for chickens.
}
\references{
IPCC. (2019). \emph{2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories}, Chapter 10: Emissions from
Livestock and Manure Management, Equation 10.21.

IPCC. (2006). \emph{2006 IPCC Guidelines for National Greenhouse Gas Inventories}, Chapter 10: Emissions from
Livestock and Manure Management, Equation 10.21.
}
