% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_simulation_core_model.R
\name{compute_fecundity_rates}
\alias{compute_fecundity_rates}
\title{Compute Daily Fecundity Rates}
\usage{
compute_fecundity_rates(parturition_rate, litter_size, birth_fraction_female)
}
\arguments{
\item{parturition_rate}{Numeric. Parturition rate (probability of an adult female to give birth).}

\item{litter_size}{Numeric. Prolificacy rate (mean number of offspring per parturition).}

\item{birth_fraction_female}{Numeric. Female birth ratio (probability that an offspring is female).}
}
\value{
A named list with two elements:
\describe{
\item{fecundity_female}{Daily number of female offspring per adult female.}
\item{fecundity_male}{Daily number of male offspring per adult female.}
}
}
\description{
Calculates the daily number of male and female offspring produced per adult female.
}
\examples{
compute_fecundity_rates(parturition_rate = 0.8, litter_size = 2, birth_fraction_female = 0.5)

}
