% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_simulation_core_model.R
\name{compute_fecundity_rates}
\alias{compute_fecundity_rates}
\title{Compute Daily Fecundity Rates}
\usage{
compute_fecundity_rates(part_rate, prolif_rate, fem_birth_ratio)
}
\arguments{
\item{part_rate}{Numeric. Parturition rate (probability of an adult female to give birth).}

\item{prolif_rate}{Numeric. Prolificacy rate (mean number of offspring per parturition).}

\item{fem_birth_ratio}{Numeric. Female birth ratio (probability that an offspring is female).}
}
\value{
A named list with two elements:
\describe{
\item{fem_fec}{Daily number of female offspring per adult female.}
\item{male_fec}{Daily number of male offspring per adult female.}
}
}
\description{
Calculates the daily number of male and female offspring produced per adult female.
}
\examples{
compute_fecundity_rates(part_rate = 0.8, prolif_rate = 2, fem_birth_ratio = 0.5)

}
