% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production_cohort_core_model.R
\name{compute_fibre_output}
\alias{compute_fibre_output}
\title{Compute Fibre Production}
\usage{
compute_fibre_output(fibre_prod, assessment_duration, size)
}
\arguments{
\item{fibre_prod}{Numeric. Fibre yield per head per year (kg/head/year).}

\item{assessment_duration}{Numeric. Length of the assessment period (days)}

\item{size}{Numeric. Herd size (heads) for the cohort.}
}
\value{
Numeric. Total fibre produced over the assessment period  by cohort (kg /cohort/assessment period).
}
\description{
Compute Fibre Production
}
