% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production_cohort_core_model.R
\name{compute_fibre_output}
\alias{compute_fibre_output}
\title{Compute Fibre Production}
\usage{
compute_fibre_output(fibre_yield, assessment_duration, size)
}
\arguments{
\item{fibre_yield}{Numeric. Fibre yield per head per day (kg/head/day).}

\item{assessment_duration}{Numeric. Number of assessment days used to annualise fibre output.}

\item{size}{Numeric. Herd size (heads) for the cohort.}
}
\value{
Numeric. Fibre production per cohort (kg/year).
}
\description{
Compute Fibre Production
}
