% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrogen_balance_core_model.R
\name{compute_nitrogen_excretion}
\alias{compute_nitrogen_excretion}
\title{Compute Daily Nitrogen Excretion}
\usage{
compute_nitrogen_excretion(animal, n_intake, n_retention)
}
\arguments{
\item{animal}{Character. Code identifying the livestock species.
Supported values include:
\itemize{
\item \code{PGS}: pigs
\item \code{CML}: camels
\item \code{CTL}: cattle
\item \code{BFL}: buffalo
\item \code{SHP}: sheep
\item \code{GTS}: goats
}}

\item{n_intake}{Numeric. Daily nitrogen intake (kg N/head/day)}

\item{n_retention}{Numeric.  Daily nitrogen retention in animal body tissues and products (e.g., growth, pregnancy, milk...) (kg N/head/day)}
}
\value{
Numeric. Daily nitrogen excretion (kg N/head/day).
}
\description{
Calculates daily nitrogen excretion per animal (kg N/head/day) as the difference between
nitrogen intake and nitrogen retention.
}
\details{
Nitrogen excretion represents the fraction of consumed nitrogen that is not
retained in animal products or body tissues and is therefore excreted in urine
and dung. This quantity forms the basis for subsequent calculations of
nitrous oxide \code{N₂O} emissions from manure management and agricultural soils.
The approach follows the IPCC 2019 guidelines.
}
\references{
IPCC. (2019). \emph{2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories}, Chapter 10: Emissions from
Livestock and Manure Management. Equation 10.31A.
}
