% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_trajectory.R
\name{differential_pathway_trajectory}
\alias{differential_pathway_trajectory}
\title{Differential pathway along trajectory}
\usage{
differential_pathway_trajectory(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=test_pathway_trajectory]{test_pathway_trajectory()}}.}
}
\value{
\code{gleam_test} object.
}
\description{
Differential pathway along trajectory
}
