% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleam_constants.R
\docType{data}
\name{gleam_cohorts}
\alias{gleam_cohorts}
\title{Valid cohort short codes}
\format{
A character vector of length 6.
}
\usage{
gleam_cohorts
}
\description{
All sex- and age-class cohort codes recognised by GLEAM.
}
\keyword{internal}
