% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleam_constants.R
\docType{data}
\name{gleam_cohorts_female}
\alias{gleam_cohorts_female}
\title{Female cohort codes}
\format{
A character vector of length 3.
}
\usage{
gleam_cohorts_female
}
\description{
Female cohort codes
}
\keyword{internal}
