% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleam_constants.R
\docType{data}
\name{gleam_cohorts_male}
\alias{gleam_cohorts_male}
\title{Male cohort codes (M-prefix)}
\format{
A character vector of length 3.
}
\usage{
gleam_cohorts_male
}
\description{
Male cohort codes (M-prefix)
}
\keyword{internal}
