% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleam_constants.R
\name{gleam_species_non_poultry}
\alias{gleam_species_non_poultry}
\title{Species with non-zero enteric CH4 and N excretion}
\format{
A character vector of length 6.
}
\usage{
gleam_species_non_poultry
}
\description{
All supported species. Used for enteric methane emission calculations and
nitrogen excretion models. (Legacy alias; equivalent to \code{gleam_species}.)
}
\keyword{internal}
