% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_in_celltype.R
\name{plot_group_in_celltype}
\alias{plot_group_in_celltype}
\title{Plot group comparison within a fixed celltype}
\usage{
plot_group_in_celltype(x, top_n = 20)
}
\arguments{
\item{x}{\code{scpathway_test} object or result table.}

\item{top_n}{Number of pathways to display.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot group comparison within a fixed celltype
}
