% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap of aggregated pathway scores}
\usage{
plot_heatmap(
  score,
  by,
  fun = c("mean", "median", "fraction"),
  threshold = 0,
  top_n = NULL,
  palette = "gleam_continuous"
)
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{by}{Grouping metadata columns.}

\item{fun}{Aggregation function.}

\item{threshold}{Threshold used by \code{fraction}.}

\item{top_n}{Optional top pathways by variance.}

\item{palette}{Continuous palette name or custom colors.}
}
\value{
A \code{ggplot} object.
}
\description{
Heatmap of aggregated pathway scores
}
