% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ridge.R
\name{plot_ridge}
\alias{plot_ridge}
\title{Ridge plot for pathway score distributions}
\usage{
plot_ridge(score, pathway, group)
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{pathway}{Pathway name.}

\item{group}{Group variable.}
}
\value{
A \code{ggplot} object.
}
\description{
Ridge plot for pathway score distributions
}
