% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin.R
\name{plot_violin}
\alias{plot_violin}
\title{Violin plot of pathway scores}
\usage{
plot_violin(score, pathway, group, celltype = NULL, trim = TRUE)
}
\arguments{
\item{score}{\code{scpathway_score} object.}

\item{pathway}{Pathway name.}

\item{group}{Group variable (metadata column name or vector).}

\item{celltype}{Optional celltype filter (single label or vector).}

\item{trim}{Passed to \code{geom_violin()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Violin plot of pathway scores
}
