% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot for differential pathways}
\usage{
plot_volcano(x, p_col = "p_adj", effect_col = "effect_size", sig_thresh = 0.05)
}
\arguments{
\item{x}{\code{scpathway_test} object or result data.frame.}

\item{p_col}{P-value column name.}

\item{effect_col}{Effect size column name.}

\item{sig_thresh}{Significance threshold for adjusted p-value.}
}
\value{
A \code{ggplot} object.
}
\description{
Volcano plot for differential pathways
}
