% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_simulation_core_model.R
\name{project_population_size}
\alias{project_population_size}
\title{Project One Year of Steady-State Population Dynamics}
\usage{
project_population_size(
  size_total,
  female_fecundity,
  male_fecundity,
  pdea,
  poff,
  g,
  growth_rate_pop,
  structure,
  share
)
}
\arguments{
\item{size_total}{Numeric. Total population size at the start of the year.}

\item{female_fecundity}{Numeric. Daily female births per adult female.}

\item{male_fecundity}{Numeric. Daily male births per adult female.}

\item{pdea}{Named numeric vector of length 10. Daily death probabilities. Must be named using:
\code{FB}, \code{FJ}, \code{FS}, \code{FA}, \code{FC}, \code{MB}, \code{MJ}, \code{MS}, \code{MA}, \code{MC}.}

\item{poff}{Named numeric vector of length 10. Daily offtake probabilities. Names must match those in \code{pdea}.}

\item{g}{Named numeric vector of length 10. Transition probabilities to next age class. Names must match those in \code{pdea}.}

\item{growth_rate_pop}{Numeric. Annual population growth rate.}

\item{structure}{Named numeric vector of length 8. Final population share in 8 classes. Must be named:
\code{FB}, \code{FJ}, \code{FS}, \code{FA}, \code{MB}, \code{MJ}, \code{MS}, \code{MA}.}

\item{share}{Named numeric vector of length 6. Final population share in 6 grouped classes. Must be named:
\code{FJ}, \code{FS}, \code{FA}, \code{MJ}, \code{MS}, \code{MA}.}
}
\value{
A named list with:
\describe{
\item{size}{Size in 6 sex-age classes at the start of the year.}
\item{size_end}{Size at year end (projected using growth rate).}
\item{size_end_exact}{Size at year end (based on full simulation over 366 days).}
\item{size_avg}{Average population size over the year.}
\item{offtake}{Total offtake over the year per sex-age class.}
}
}
\description{
Simulates one year of population dynamics under steady-state assumptions using demographic parameters
and returns population size statistics and offtake results.
}
