% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_directemissions_entric.R
\name{run_directemissions_enteric}
\alias{run_directemissions_enteric}
\title{Run Enteric Methane Direct Emissions (Internal)}
\usage{
run_directemissions_enteric(data)
}
\arguments{
\item{data}{A \code{data.table} with cohort-level nutritional and demographic inputs.}
}
\value{
The same \code{data.table} with new columns \code{ym} and \code{ch4_enteric}.

#' IPCC. (2019). \emph{2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories}, Chapter 10: Emissions from
Livestock and Manure Management, Equation 10.21.

IPCC. (2006). \emph{2006 IPCC Guidelines for National Greenhouse Gas Inventories}, Chapter 10: Emissions from
Livestock and Manure Management, Equation 10.21.
}
\description{
Computes daily enteric methane emissions (kg CH₄/head/day) for each
cohort record by applying species-, cohort- and diet- specific methane conversion factors (YM),
using a Tier 2 approach (IPCC 2006, 2019)
}
\details{
This function is intended for internal workflows and does not perform any file I/O.

It adds two columns:
\itemize{
\item \code{ym}: Methane conversion factor (\% of gross energy intake converted to CH4).
\item \code{ch4_enteric}: Daily enteric methane emissions (kg CH4/head/day).
}

Input data must at minimum include the following columns:
\itemize{
\item \code{Animal_short}: Character. Species code  (e.g., \code{PGS}, \code{CML}, \code{CTL}, \code{BFL}, \code{SHP}, \code{GTS}).
\item \code{cohort}: Character. Cohort code (e.g., \code{FA}, \code{FS}, \code{FJ}, \code{MA}, \code{MS},\code{MJ}).
\item \code{diet_dig}: Numeric. Average digestibility of the the feed ration, expressed as ratio of digestible to gross energy content (fraction)
\item \code{diet_ge}: Numeric. Average gross energy content of the diet (MJ/kg DM).
\item \code{dmi}: Numeric. Daily dry matter intake of feed (kg DM/head/day).
}
}
\examples{
\dontrun{
# Load example input from the package and run the simulation
input_path <- system.file("extdata/GLEAM_input_directemissions_enteric.csv", package = "gleam")
dt <- data.table::fread(input_path)
result <- run_directemissions_enteric(dt)
head(result[, .(Animal_short, cohort, ym, ch4_enteric)])
}

}
\keyword{internal}
