% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_run.R
\name{run_scpathway}
\alias{run_scpathway}
\title{End-to-end scPathway workflow (deprecated alias)}
\usage{
run_scpathway(...)
}
\value{
List containing score object and comparison result.
}
\description{
End-to-end scPathway workflow (deprecated alias)
}
