% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_expr}
\alias{toy_expr}
\title{Toy expression dataset}
\format{
A list with components \code{expr} and \code{meta}.
}
\source{
Simulated toy data
}
\usage{
toy_expr
}
\description{
A list with expression matrix and metadata for tests/examples.
}
\keyword{datasets}
