% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_trajectory.R
\name{trajectory_table}
\alias{trajectory_table}
\title{Build trajectory table from score object}
\usage{
trajectory_table(
  score,
  pseudotime = NULL,
  lineage = NULL,
  embeddings = NULL,
  backend = c("auto", "internal", "monocle3", "monocle", "slingshot")
)
}
\arguments{
\item{score}{\code{gleam_score} object.}

\item{pseudotime}{Pseudotime source.}

\item{lineage}{Lineage source.}

\item{embeddings}{Optional embedding matrix.}

\item{backend}{Trajectory backend. \code{auto} detects from provided inputs.}
}
\value{
Data.frame.
}
\description{
Build trajectory table from score object
}
